### 2024-09-03 #7170 @editorialbot
[PRE REVIEW]: tempodisco: an R package for temporal discounting
**Submitting author:** @kinleyid (<a href="http://orcid.org/0000-0003-2057-9606">Isaac Kinley</a>)
**Repository:** https://github.com/kinleyid/tempodisco/
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
