### 2024-09-03 c3b6bb0
Update license and ORCID
