+++ paper/paper.bib
@@ -6,6 +6,7 @@
   number={3},
   pages={264},
   year={2008},
+  doi={10.1037/1064-1297.16.3.264},
   publisher={American Psychological Association}
 }
 
@@ -17,6 +18,7 @@
   number={1},
   pages={78},
   year={1999},
+  doi={10.1037//0096-3445.128.1.78},
   publisher={American Psychological Association}
 }
 
@@ -27,6 +29,7 @@
   volume={36},
   pages={41--51},
   year={2004},
+  doi={10.3758/bf03195548},
   publisher={Springer}
 }
 
@@ -38,6 +41,7 @@
   number={1},
   pages={218--233},
   year={2015},
+  doi={10.1002/jeab.128},
   publisher={Wiley Online Library}
 }
 
@@ -49,6 +53,7 @@
   number={2},
   pages={235--243},
   year={2001},
+  doi={10.1901/jeab.2001.76-235},
   publisher={Wiley Online Library}
 }
 
@@ -60,6 +65,7 @@
   number={1-2},
   pages={169--172},
   year={2008},
+  doi={10.1016/j.drugalcdep.2007.12.011},
   publisher={Elsevier}
 }
 
@@ -69,6 +75,7 @@
   journal={Journal of visualized experiments: JoVE},
   number={107},
   year={2016},
+  doi={10.3791/53584},
   publisher={MyJoVE Corporation}
 }
 
@@ -77,6 +84,7 @@
   author={Gelino, Brett W and Strickland, Justin C and Johnson, Matthew W},
   journal={Journal of the Experimental Analysis of Behavior},
   year={2024},
+  doi={10.1002/jeab.4200},
   publisher={Wiley Online Library}
 }
 
@@ -88,5 +96,6 @@
   number={11},
   pages={1799--1806},
   year={2021},
+  doi={10.1017/S0033291721002282},
   publisher={Cambridge University Press}
 }
