### 2024-09-03 #7170 @kinleyid
Hello, I'd like to add something I forgot to mention in the submission note---in addition to being under review, the paper that used an early version of this software also exists as a preprint that contains a link to said early version: https://doi.org/10.31234/osf.io/y2fdh
