+++ README-old.md
@@ -1,112 +1,0 @@
-
-# tempodisco
-
-[![R-CMD-check](https://github.com/kinleyid/tempodisco/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/kinleyid/tempodisco/actions/workflows/R-CMD-check.yaml)
-[![codecov](https://codecov.io/github/kinleyid/tempodisco/graph/badge.svg?token=CCQXS3SNGB)](https://codecov.io/github/kinleyid/tempodisco)
-
-An R package for working with temporal discounting data.
-
-## Installation
-```
-require(devtools)
-install_github("kinleyid/tempodisco");
-```
-
-## Description
-This package allows you to fit 2 types of models: **binary choice models** (BCMs), which are probabilistic models of individual binary choices, and **indifference point models** (IPMs), which are models of indifference points. In both cases, this package tests 7 candidate discount functions and identifies the best fit to an individual's data (the one yielding the lowest BIC):
-
-| Name | Functional form |
-|------|-----------------|
-| Exponential (Samuelson, 1937) |	$f(t; k) = e^{-k t}$ |
-| Scaled exponential (beta-delta; Laibson, 1997) | $f(t; k, w) = w e^{-k t}$ |
-| Nonlinear-time exponential (Ebert & Prelec, 2007) | $f(t; k, s) = e^{-k t^s}$ |
-| Dual-systems exponential (Ven den Bos & McClure, 2013) | $f(t; k_1, k_2, w) = w e^{-k_1 t} + (1 - w) e^{-k_2 t}$ |
-| Inverse q-exponential (Green & Myerson, 2004) | $f(t; k, s) = \frac{1}{(1 + k t)^s}$ |
-| Hyperbolic (Mazur, 1987) | $f(t; k) = \frac{1}{1 + kt}$ |
-| Nonlinear-time hyperbolic (Rachlin, 2006) | $f(t; k, s) = \frac{1}{1 + k t^s}$ |
-
-Additionally, 2 other discount functions are tested: a "model-free" function in which each indifference point is a different parameter, and a "noise" function that only specifies an intercept (i.e., $f(t; c) = c$).
-
-## Example usage
-
-### Fitting binary choice models: `td_bcnm` and `td_bclm`
-
-To fit a binary choice model, we need data from a single participant formatted as follows:
-
-```R
-data("td_bc_single_ptpt")
-```
-
-| val_imm | val_del | del | imm_chosen |
-|--|--|--|--|
-|112 |    187 |   30.4167 |      FALSE |
-|37 |     186 | 3652.5000  |      TRUE |
-| ... | ... | ... | ... |
-
-Here, each row corresponds to a difference decision. *val_imm* specifies the value of the immediate reward, *val_del* specifies the value of the delayed reward, *del* speifies the delay of the delayed reward, and *imm_chosen* specifies whether the participant selected the immediate reward. There can be additional columns (e.g., containing reaction times or a participant identifier), but at least these ones are required.
-
-From here, we can fit a binary choice model:
-
-```R
-mod <- td_bcnm(td_bc_single_ptpt)
-```
-
-By default, all of the discount functions above are tested. If we are interested in only a subset of these, we can specify them as follows:
-
-```R
-mod <- td_bcnm(td_bc_single_ptpt, discount_function = c('hyperbolic', 'exponential'))
-```
-
[…]
