### 2024-09-04 a9e59d6
Re-add badges
