### 2024-09-15 91f6218
Allowing multiple user-defined discount functions
