### 2024-09-15 bd35a2f
Allow fitting single model-free indifference point
