### 2024-09-16 088f8bd
Add validation functions
