+++ man/attention_checks.Rd
@@ -1,0 +1,26 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/validation-functions.R
+\name{attention_checks}
+\alias{attention_checks}
+\title{Run attention checks}
+\usage{
+attention_checks(data, warn = F)
+}
+\arguments{
+\item{data}{A \code{data.frame} with columns \code{val_imm}, \code{val_del} and \code{del_chosen}, representing data from a single participant.}
+
+\item{warn}{Logical: give a warning for failed attention checks?}
+}
+\value{
+Named vector counting the number of times the participant chose an immediate reward of 0 (\code{imm_0}) or chose a delayed reward equal in face value to an immediate reward (\code{imm_eq_del}).
+}
+\description{
+Check whether participants failed attention checks, either choosing an immediate reward of 0 or choosing a delayed reward equal in face value to an immediate reward. If the participant was never offered either choice, a warning is given.
+}
+\examples{
+\dontrun{
+# On a model
+data("td_bc_single_ptpt")
+attention_checks(td_bc_single_ptpt)
+}
+}
