+++ NAMESPACE
@@ -20,6 +20,9 @@ S3method(residuals,td_ipm)
 export(AUC)
 export(ED50)
 export(adj_amt_indiffs)
+export(attention_checks)
+export(discount_function)
+export(invariance_checks)
 export(kirby_score)
 export(nonsys)
 export(td_bclm)
