### 2024-09-25 eeffd60
Compute consistency scores from MCQ scoring
