+++ man/kirby_consistency.Rd
@@ -1,0 +1,25 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/utils.R
+\name{kirby_consistency}
+\alias{kirby_consistency}
+\title{Compute consistency score}
+\usage{
+kirby_consistency(data, discount_function = c("hyperbolic", "exponential"))
+}
+\arguments{
+\item{data}{Responses to score.}
+
+\item{discount_function}{Should \eqn{k} values be computed according to the hyperbolic or exponential discount function? The original method uses the hyperbolic, but in principle the exponential is also possible.}
+}
+\value{
+A consistency score
+}
+\description{
+Compute the consistency score per the method of \href{https://doi.org/10.1037//0096-3445.128.1.78}{Kirby et al. (1999)}. This is described in detail in \href{https://doi.org/10.1007/s40614-016-0070-9}{Kaplan et al. (2016)}, where it's suggested that a consistency score below 0.75 might be a sign of inattentive responding.
+}
+\examples{
+\dontrun{
+data("td_bc_single_ptpt")
+mod <- kirby_consistency(td_bc_single_ptpt)
+}
+}
