+++ NAMESPACE
@@ -23,6 +23,7 @@ export(adj_amt_indiffs)
 export(attention_checks)
 export(discount_function)
 export(invariance_checks)
+export(kirby_consistency)
 export(kirby_score)
 export(nonsys)
 export(td_bclm)
