### 2024-09-25 #7170 @kinleyid
Hello, just wanted to suggest some reviewers with R + decision making research expertise:
- [SeHellmann](https://github.com/SeHellmann)
- [bkrayfield](https://github.com/bkrayfield)
- [haoxue-fan](https://github.com/haoxue-fan)
