### 2024-09-29 c142ed7
Update terminology: 'intercept-only' -> 'constant'
