### 2024-09-29 4e90016
Add deviance computation
