### 2024-09-29 ea48b60
Optionally compute proportion failed attn checks
