+++ R/validation-functions.R
@@ -13,6 +13,7 @@ require_columns <- function(data, columns) {
 #' Check whether participants failed attention checks, either choosing an immediate reward of 0 or choosing a delayed reward equal in face value to an immediate reward. If the participant was never offered either choice, a warning is given.
 #' @param data A \code{data.frame} with columns \code{val_imm}, \code{val_del} and \code{del_chosen}, representing data from a single participant.
 #' @param warn Logical: give a warning for failed attention checks?
+#' @param ppn Logical: return proportions of attention checks participant failed, versus absolute numbers?
 #' @returns Named vector counting the number of times the participant chose an immediate reward of 0 (\code{imm_0}) or chose a delayed reward equal in face value to an immediate reward (\code{del_eq_imm}).
 #' @examples
 #' \dontrun{
@@ -21,7 +22,7 @@ require_columns <- function(data, columns) {
 #' attention_checks(td_bc_single_ptpt)
 #' }
 #' @export
-attention_checks <- function(data, warn = F) {
+attention_checks <- function(data, warn = F, ppn = F) {
   
   require_columns(data, c('val_imm', 'val_del', 'imm_chosen'))
   
