### 2024-09-29 411063a
Bug fix---compute proportion attn checks
