### 2024-11-06 28013fc
Remove undocumented function
To be added in a future release
