### 2024-11-07 dcfe5f2
Add support for drift diffusion models
Still need to add tests
