+++ R/generics.R
@@ -33,6 +33,18 @@ print.td_ipm <- function(x, ...) {
 }
 
 #' @export
+print.td_ddm <- function(x, ...) {
+  cat(sprintf('\nTemporal discounting drift diffusion model\n\n'))
+  cat(sprintf('Discount function: %s\nCoefficients:\n\n', x$config$discount_function$name))
+  print(coef(x))
+  cat(sprintf('\n"%s" transform applied to drift rates.\n', x$config$drift_transform$name))
+  cat(sprintf('\nED50: %s\n', ED50(x)))
+  cat(sprintf('AUC: %s\n', AUC(x, verbose = F)))
+  cat(sprintf('BIC: %s\n', BIC(x)))
+}
+
+
+#' @export
 print.td_fn <- function(x, ...) {
   obj <- x
   cat(sprintf('\n"%s" temporal discounting function\n\n', obj$name))
@@ -49,12 +61,25 @@ print.td_fn <- function(x, ...) {
   }
 }
 
+predict_indiffs <- function(object, newdata) {
+  
+  # Predict indiffs for an object that has:
+  #   object$config$discount_function$fn
+  #   coef(object)
+  
+  indiff_func <- object$config$discount_function$fn
+  indiffs <- indiff_func(newdata, coef(object))
+  names(indiffs) <- NULL
+  return(indiffs)
+  
+}
+
 #' Model Predictions
 #'
 #' Generate predictions from a temporal discounting binary choice model
 #' @param object A temporal discounting binary choice model. See \code{td_bcnm}.
 #' @param newdata Optionally, a data frame to use for prediction. If omitted, the data used to fit the model will be used for prediction.
-#' @param type The type of prediction required. As in predict.glm, \code{"link"} (default) and \code{"response"} give predictions on the scales of the linear predictors and response variable, respectively. \code{"indiff"} gives predicted indifference points. In this case, \code{newdata} needs only a \code{del} column.
+#' @param type The type of prediction required. As in predict.glm, \code{"link"} (default) and \code{"response"} give predictions on the scales of the linear predictors and response variable, respectively. \code{"indiff"} gives predicted indifference points. For predicting indifference points, \code{newdata} needs only a \code{del} column.
 #' @param ... Additional arguments currently not used.
 #' @return A vector of predictions
 #' @examples
@@ -86,10 +111,12 @@ predict.td_bcnm <- function(object, newdata = NULL, type = c('link', 'response',
     
   } else if (type == 'indiff') {
     
-    indiff_func <- object$config$discount_function$fn
-    indiffs <- indiff_func(newdata, coef(object))
-    names(indiffs) <- NULL
-    return(indiffs)
+    return(predict_indiffs(object, newdata))
+    
+    # indiff_func <- object$config$discount_function$fn
+    # indiffs <- indiff_func(newdata, coef(object))
+    # names(indiffs) <- NULL
+    # return(indiffs)
     
   }
 }
@@ -118,7 +145,7 @@ predict.td_bclm <- function(object, newdata = NULL, type = c('indiff', 'link', '
   type <- match.arg(type)
   if (type == 'indiff') {
     
-    return(predict.td_bcnm(object, newdata = newdata, type = type))
+    return(predict_indiffs(object, newdata))
     
   } else {
     
@@ -171,6 +198,65 @@ predict.td_ipm <- function(object, newdata = NULL, type = c('indiff', 'response'
   
 }
 
+#' Model Predictions
+#'
+#' Generate predictions from a temporal discounting drift diffusion model
+#' @param object A temporal discounting drift diffusion model. See \link{td_ddm}.
[…]
