+++ R/td_ddm.R
@@ -1,0 +1,278 @@
+
+# TODO: Test multiple candidate discount functions.
+# TODO: Use the same terminology (get_linpred_func_bcnm) for td_bcnm as here
+# TODO: Update td_bcnm to use get_candidate_discount_functions
+# TODO: Custom criterion functions, both here and td_bcnm
+# TODO: Write tests
+
+# A note on conventions:
+# UPPER boundary represents immediate reward
+# LOWER boundary represents delayed reward
+# Densities are by default computed at upper boundary
+# To compute at lower boundary, use -v and (1 - w)
+# Subjective value difference should be positive if favouring imm., negative if favouring del.
+# 
+# v is a drift rate *multiplier*. I.e., not the drift rate itself
+# alpha = boundary separation
+# tau = non-decision-time
+# beta = bias
+
+
+#' Temporal discounting drift diffusion model
+#'
+#' Fit a drift diffusion model for a single subject using maximum likelihood estimation.
+#' @param data A data frame with columns \code{val_imm} and \code{val_del} for the values of the immediate and delayed rewards, \code{del} for the delays, \code{imm_chosen} (Boolean) for whether the immediate rewards were chosen, and \code{rt} for the reaction times (in seconds). Other columns can also be present but will be ignored.
+#' @param discount_function A string specifying the name of the discount functions to use, or an object of class \code{td_fn} (used for creating custom discount functions), or a list of objects of class \code{td_fn}.
+#' @param v_par_starts A vector of starting values to try for the "v" parameter (drift rate multiplier) during optimization.
+#' @param beta_par_starts A vector of starting values to try for the "beta" parameter (bias) during optimization.
+#' @param alpha_par_starts A vector of starting values to try for the "alpha" parameter (boundary separation) during optimization.
+#' @param tau_par_starts A vector of starting values to try for the "tau" parameter (non-decision time) during optimization.
+#' @param drift_transform A transform to apply to drift rates. Either \code{"none"} (no transform), \code{"sigmoid"} (sigmoidal transform described by \href{https://doi.org/10.1371/journal.pcbi.1007615}{Peters & D'Esposito, 2020} and \href{https://doi.org/10.3758/s13423-018-1554-2}{Fontanesi et al., 2019}), or \code{"bias-correct"} (experimental; see note below).
+#' @param optim_args Additional arguments to pass to \code{optim()}. Default is \code{list(silent = T)}.
+#' @param silent Boolean (true by default). The call to \code{optim()} occurs within a \code{try()} wrapper. The value of \code{silent} is passed along to \code{try()}.
+#' @param na.action Action to take when data contains \code{NA} values. Default is \code{na.omit}.
+#' @return An object of class \code{td_bcnm} with components \code{data} (containing the data used for fitting), \code{config} (containing the internal configuration of the model, including the \code{discount_function}), and \code{optim} (the output of \code{optim()}).
+#' @note
[…]
