+++ R/data-docs.R
@@ -10,7 +10,7 @@ NULL
 
 #' Binary choice data for a study
 #' 
-#' Data from 421 participants, who each made 70 binary choices. Along with the columns required by \code{td_bcnm}, the reaction time (\code{rt}) is recorded. Participants are identified by the alphnumeric code in the \code{id} column.
+#' Data from 421 participants, who each made 70 binary choices. Along with the columns required by \code{td_bcnm}, the reaction time (\code{rt}) in seconds is recorded. Participants are identified by the alphnumeric code in the \code{id} column.
 #' @name td_bc_study
 #' @docType data
 #' @author Isaac Kinley \email{isaac.kinley@gmail.com}
