+++ vignettes/analyzing-study.Rmd
@@ -33,6 +33,7 @@ As a first step, we might want to remove cases of non-systematic discounting:
 # Check for nonsystematic discounting
 is_nonsys <- by(study, study$id, function(ptpt) { # Stratify by participant ID
   mod <- td_bcnm(ptpt, discount_function = 'model-free') # Fit a "model-free" discount function to the participant-level data
+  print(mod)
   any(nonsys(mod)) # Check whether either non-systematic criterion is met
 }, simplify = T)
 # Keep data from systematic discounters

[…]
