### 2024-11-07 2091b8a
Require "vctrs" package in tests
