+++ NAMESPACE
@@ -7,6 +7,7 @@ S3method(coef,td_ipm)
 S3method(deviance,td_bcnm)
 S3method(deviance,td_ddm)
 S3method(fitted,td_bcnm)
+S3method(fitted,td_ddm)
 S3method(fitted,td_ipm)
 S3method(logLik,td_bcnm)
 S3method(logLik,td_ddm)
@@ -40,6 +41,7 @@ export(td_ipm)
 export(wileyto_score)
 importFrom(RWiener,dwiener)
 importFrom(grDevices,rgb)
+importFrom(graphics,axis)
 importFrom(graphics,lines)
 importFrom(graphics,points)
 importFrom(graphics,title)
@@ -52,8 +54,10 @@ importFrom(stats,coef)
 importFrom(stats,fitted)
 importFrom(stats,glm)
 importFrom(stats,integrate)
+importFrom(stats,median)
 importFrom(stats,na.omit)
 importFrom(stats,optim)
+importFrom(stats,plogis)
 importFrom(stats,predict)
 importFrom(stats,predict.glm)
 importFrom(stats,qlogis)
