### 2024-11-07 ea5033a
Remove dependency on vctrs?
