+++ tests/testthat/test-td_bcnm.R
@@ -1,6 +1,4 @@
 
-require(vctrs)
-
 # Get data
 data("td_bc_single_ptpt")
 df <- td_bc_single_ptpt
@@ -61,9 +59,9 @@ while (arg_combo_idx <= nrow(arg_combos)) {
     expect_type(AIC(mod, k = 5), 'double')
     expect_type(logLik(mod), 'double')
     expect_named(coef(mod))
-    expect_vector(residuals(mod, type = 'deviance'), size = nrow(df))
-    expect_vector(residuals(mod, type = 'pearson'), size = nrow(df))
-    expect_vector(residuals(mod, type = 'response'), size = nrow(df))
+    expect_length(residuals(mod, type = 'deviance'), n = nrow(df))
+    expect_length(residuals(mod, type = 'pearson'), n = nrow(df))
+    expect_length(residuals(mod, type = 'response'), n = nrow(df))
   })
   
   pdf(NULL) # Don't actually produce plots
@@ -79,10 +77,10 @@ while (arg_combo_idx <= nrow(arg_combos)) {
   
   # prediction
   test_that('predictions', {
-    expect_vector(fitted(mod), ptype = numeric(0), size = nrow(df))
-    expect_vector(predict(mod, type = 'link'), ptype = numeric(0), size = nrow(df))
-    expect_vector(predict(mod, type = 'response'), ptype = numeric(0), size = nrow(df))
-    expect_vector(predict(mod, newdata = data.frame(del = 0:1000), type = 'indiff'), ptype = numeric(0), size = 1001)
+    expect_length(fitted(mod), n = nrow(df))
+    expect_length(predict(mod, type = 'link'), n = nrow(df))
+    expect_length(predict(mod, type = 'response'), n = nrow(df))
+    expect_length(predict(mod, newdata = data.frame(del = 0:1000), type = 'indiff'), n = 1001)
   })
 }
 
