### 2024-11-07 b08b61a
Remove vctrs dependency
