+++ README.Rmd
@@ -120,6 +120,7 @@ print(ddm)
 ```
 
 Following [Peters & D'Esposito (2020)](https://doi.org/10.1371/journal.pcbi.1007615), we can apply a sigmoidal transform to the drift rate $\delta$ to improve model fit using the argument `drift_transform = "sigmoid"`:
+
 $$\delta' = v_\text{max} \left(\frac{2}{1 + \exp\{-\delta\}} - 1\right)$$
 
 ```{r}
@@ -140,8 +141,6 @@ We can use the model to predict reaction times and compare them to the actual da
 pred_rts <- predict(ddm_sig, type = 'rt')
 cor.test(pred_rts, td_bc_single_ptpt$rt)
 ```
-
-This implementation follows the model described in [Peters & D'Esposito (2020)](https://doi.org/10.1371/journal.pcbi.1007615); the drift rate is based on the difference in subjective value between the immediate and delayed rewards, where the subjective value of 
 
 ### The "model-free" discount function
 
