### 2024-11-08 eea493e
Elaborate on DDM models and fix vctrs dependency
