+++ tests/testthat/test-td_ddm.R
@@ -1,6 +1,4 @@
 
-require(vctrs)
-
 # Get data
 data("td_bc_single_ptpt")
 df <- td_bc_single_ptpt
