+++ vignettes/drift-diffusion-models.Rmd
@@ -18,7 +18,23 @@ knitr::opts_chunk$set(
 library(tempodisco)
 ```
 
-Drift diffusion models (DDMs; [Ratcliff, 1978](https://doi.org/10.1037/0033-295X.85.2.59)) model not only binary choices but also response times (RTs), and can be applied to intertemporal choice data ([Peters & D'Esposito, 2020](https://doi.org/10.1371/journal.pcbi.1007615)). Often, extreme RTs are first excluded, using either absolute or relative cutoffs. Here, we will exclude the fastest 2.5% and slowest 2.5% of responses.
+## The drift diffusion model
+
+Drift diffusion models (DDMs; [Ratcliff, 1978](https://doi.org/10.1037/0033-295X.85.2.59)) predict not only binary choices but also response times (RTs), and can be applied to intertemporal choice data ([Peters & D'Esposito, 2020](https://doi.org/10.1371/journal.pcbi.1007615)). DDMs model decision making as a process of noisy evidence accumulation in which the evidence $X$ begins at some value $\beta$ and evolves according to the stochastic differential equation
+
+$$\frac{d}{dt}X \sim \nu + \omega$$
+
+where $\delta$ is the drift rate, and $\omega$ is normally distributed noise with 0 mean and unit variance. Evidence begins accumulating after a non-decision period $\tau$ and is complete once $X$ reaches either 0 (meaning the delayed reward is chosen) or a threshold $\alpha$ (meaning the immediate reward is chosen). Thus, the model captures speed-accuracy tradeoffs: as $\alpha$ increases, decisions will be slower but more "accurate" (i.e., reflective of the decision maker's relative preferences).
+
+Following [Peters & D'Esposito (2020)](https://doi.org/10.1371/journal.pcbi.1007615), the drift rate is computed by multiplying the difference in subjective values by a scale factor $v$:
+
+$$\delta = v\big( V_\text{imm} - V_\text{del}f(t;\mathbf{p})\big)$$
+
+where $V_\text{imm}$ is the value of the immediate reward, $V_\text{del}$ is the value of the delayed reward, and $f(t;\mathbf{p})$ is a discount function parameterized by $\mathbf{p}$ and evaluated at the delay of the delayed reward $t$ (i.e., $t$ in this equation does not refer to time within a single decision trial as in the previous equation).
+
+## Fitting drift diffusion models
+
+Often, extreme RTs are first excluded, using either absolute or relative cutoffs. Here, we will exclude the fastest 2.5% and slowest 2.5% of responses.
 
 ```{r}
 data("td_bc_single_ptpt")
@@ -37,11 +53,13 @@ ddm <- td_ddm(td_bc_single_ptpt, discount_function = 'exponential',
 print(ddm)
 ```
 
-As with other model classes (`td_bcnm`, `td_bclm`), we can plot the resulting discount function
+As we can see, parameters are estimated not only for the discount function ($k$), but also for the DDM ($v$, $\beta$, $\alpha$ and $\tau$). These can be extracted using `coef()` to collect for group-level analysis (see the vignette ["Analyzing data from multiple participants"](analyzling-study.html)). As with other model classes (`td_bcnm`, `td_bclm`), we can plot the resulting discount function
 
 ```{r}
 plot(ddm, log = 'x', verbose = F)
 ```
+
+Here the gaps correspond to very fast or slow decisions that we excluded.
 
 Moreover, because DDMs model RTs, we can compare predicted RTs to the actual data:
 
@@ -49,6 +67,8 @@ Moreover, because DDMs model RTs, we can compare predicted RTs to the actual dat
 predicted_rts <- predict(ddm, type = 'rt')
[…]
