### 2024-11-08 dbb27c4
Adding plots for DDMs
