+++ README.Rmd
@@ -142,6 +142,12 @@ pred_rts <- predict(ddm_sig, type = 'rt')
 cor.test(pred_rts, td_bc_single_ptpt$rt)
 ```
 
+We can also plot reaction times against the model's predictions:
+
+```{r}
+plot(ddm_sig, type = 'rt', confint = 0.8)
+```
+
 ### The "model-free" discount function
 
 In addition to the discount functions listed above, we can fit a "model-free" discount function to the data, meaning we fit each indifference point independently. This enables us to, first, test whether a participant exhibits non-systematic discounting according to the Johnson & Bickel criteria:
