+++ R/td_ddm.R
@@ -1,9 +1,7 @@
 
-# TODO: Test multiple candidate discount functions.
 # TODO: Use the same terminology (get_linpred_func_bcnm) for td_bcnm as here
 # TODO: Update td_bcnm to use get_candidate_discount_functions
 # TODO: Custom criterion functions, both here and td_bcnm
-# TODO: Write tests
 
 # A note on conventions:
 # UPPER boundary represents immediate reward
