+++ man/vcov.td_um.Rd
@@ -1,0 +1,19 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/generics.R
+\name{vcov.td_um}
+\alias{vcov.td_um}
+\title{Calculate variance-vovariance matrix for a teporal discounting model}
+\usage{
+\method{vcov}{td_um}(object, ...)
+}
+\arguments{
+\item{object}{Model for which to estimate variance-covariance matrix.}
+
+\item{...}{Additional arguments not currently used}
+}
+\description{
+Returns a variance-covariance matrix of the parameters of a temporal discounting model.
+}
+\note{
+This method only works for models where parameter fitting involved a call to optim(), which is not the case for methods such as \link{kirby_score}.
+}
