+++ man/summary.td_um.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/generics.R
+\name{summary.td_um}
+\alias{summary.td_um}
+\title{Summary of temporal discounting model}
+\usage{
+\method{summary}{td_um}(object, ...)
+}
+\description{
+Summary of temporal discounting model
+}
