+++ NAMESPACE
@@ -17,6 +17,7 @@ S3method(predict,td_bclm)
 S3method(predict,td_bcnm)
 S3method(predict,td_ddm)
 S3method(predict,td_ipm)
+S3method(print,summary.td_um)
 S3method(print,td_bclm)
 S3method(print,td_bcnm)
 S3method(print,td_ddm)
@@ -24,6 +25,8 @@ S3method(print,td_fn)
 S3method(print,td_ipm)
 S3method(residuals,td_bcnm)
 S3method(residuals,td_ipm)
+S3method(summary,td_um)
+S3method(vcov,td_um)
 export(AUC)
 export(ED50)
 export(adj_amt_indiffs)
