### 2024-11-08 bf96939
Create most_consistent_indiffs
