+++ R/tempodisco-package.R
@@ -1,5 +1,5 @@
 
-#' @importFrom stats optim predict qlogis residuals integrate coef BIC AIC glm binomial fitted approx predict.glm na.omit
+#' @importFrom stats optim predict qlogis residuals integrate coef BIC AIC glm binomial fitted approx predict.glm na.omit filter
 #' @importFrom graphics lines points title
 #' @importFrom methods is
 #' @importFrom grDevices rgb
