+++ R/tempodisco-package.R
@@ -1,6 +1,6 @@
 
-#' @importFrom stats optim predict qlogis residuals integrate coef BIC AIC glm binomial fitted approx predict.glm na.omit filter
-#' @importFrom graphics lines points title
+#' @importFrom stats optim predict qlogis plogis residuals integrate coef BIC AIC glm binomial fitted approx predict.glm na.omit median filter
+#' @importFrom graphics lines points title axis
 #' @importFrom methods is
 #' @importFrom grDevices rgb
 NULL
