### 2024-11-09 e3b8108
Faster with grid search optimization
