+++ R/generics.R
@@ -491,6 +491,7 @@ deviance.td_ddm <- function(mod) return(-2*logLik.td_ddm(mod))
 #' @param val_del Plots data for a particular delayed value
 #' @param legend Logical: display a legend? Only relevant for \code{type = 'summary'} and \code{type = 'rt'}
 #' @param p_lines Numerical vector. When \code{type = 'summary'} the discount curve, where the probability of selecting the immediate reward is 0.5, is plotted. \code{p_lines} allows you to specify other probabilities for which similar lines should be plotted (only applicable for probabilistic models, e.g. \code{td_bcnm}, \code{td_bclm} and \code{td_ddm}).
+#' @param p_tol Numerical vector. When \code{type = 'summary'} the discount curve, where the probability of selecting the immediate reward is 0.5, is plotted. \code{p_lines} allows you to specify other probabilities for which similar lines should be plotted (only applicable for probabilistic models, e.g. \code{td_bcnm}, \code{td_bclm} and \code{td_ddm}).
 #' @param verbose Whether to print info about, e.g., setting del = ED50 when \code{type = 'endpoints'}
 #' @param confint When \code{type = 'rt'}, what confidence interval should be plotted for RTs? Default is 0.95 (95\% confidence interval)
 #' @param ... Additional arguments to \code{plot()}
@@ -506,6 +507,7 @@ plot.td_um <- function(x,
                        type = c('summary', 'endpoints', 'link', 'rt'),
                        legend = T,
                        p_lines = NULL,
+                       p_tol = 0.001,
                        verbose = T,
                        del = NULL,
                        val_del = NULL,
@@ -543,28 +545,36 @@ plot.td_um <- function(x,
     # Plot indifference curve
     lines(pred_indiffs ~ plotting_delays)
     
-    # Get inverse of function that computes probability of selecting immediate reward
-    pimm_func <- function(data) predict(x, data, type = 'response')
-    # pimm_func <- get_pimm_func(x)
-    inv_pimm_func <- function(p, del) {
-      df <- data.frame(data.frame(val_imm = NA, val_del = val_del, del = del))
-      f <- function(val_imm) {
-        df$val_imm <- val_imm
-        p - pimm_func(df)
+    if (!is.null(p_lines)) {
+      # Plot curves for other probabilities
+      
+      # Because of the overhead of individual calls to predict(), exhaustive grid search
+      # is faster than uniroot() or similar to invert predict()
+      
+      # Call predict() once on a big grid
+      val_imm_cands <- seq(0, val_del, length.out = ceiling(1/p_tol + 1))
+      grid <- data.frame(val_del = val_del,
+                         del = rep(plotting_delays, each = length(val_imm_cands)),
+                         val_imm = rep(val_imm_cands, times = length(plotting_delays)))
+      grid$p <- predict(x, grid, type = 'response')
+      
+      # Split the grid by delay
+      # Using split() with a numerical index is faster than calling tapply() or similar
+      split_idx <- rep(1:length(plotting_delays), each = length(val_imm_cands))
+      subgrid_list <- split(newdata, split_idx)
+      for (p in p_lines) {
+        # Get the val_imm producing (close to) the desired p at each delay
+        val_imm <- sapply(subgrid_list, function(subgrid) {
+          if (max(subgrid$p) < p | min(subgrid$p) > p) {
+            return(NA)
+          } else {
+            return(subgrid$val_imm[which.min((subgrid$p - p)**2)])
+          }
+        })
+        # Plot
+        y <- val_imm / val_del
+        lines(y ~ plotting_delays, lty = 'dashed')
       }
-      if (f(0) <= 0 | f(val_del) >= 0) {
[…]
