### 2024-11-09 3602296
Remove unused internal function
