### 2024-11-09 f7e72d1
Add tests for experimental indiff scoring
