+++ README.Rmd
@@ -20,7 +20,7 @@ knitr::opts_chunk$set(
 [![codecov](https://codecov.io/github/kinleyid/tempodisco/graph/badge.svg?token=CCQXS3SNGB)](https://codecov.io/github/kinleyid/tempodisco)
 <!-- badges: end -->
 
-The goal of tempodisco is to provide easy access to common methods for working with temporal discounting data.
+`tempodisco` is an R package for working with delay discounting data.
 
 ## Installation
 
@@ -48,7 +48,7 @@ indiff_data <- adj_amt_indiffs(adj_amt_sim)
 head(indiff_data)
 ```
 
-This returns a data frame with a column for each delay and a column for the corresponding indifference points. The function `td_ipm` can then be used to identify the best-fitting discount function (according to the Bayesian information criterion) from any subset of the following options:
+This returns a data frame containing the delays and corresponding indifference points. The function `td_ipm` can then be used to identify the best-fitting discount function (according to the Bayesian information criterion) from any subset of the following options:
 
 ```{r child="man/fragments/predefined-discount-functions.Rmd"}
 ```
@@ -108,7 +108,7 @@ plot(mod, log = 'x', verbose = F, p_lines = c(0.05, 0.95))
 
 ### Drift diffusion models
 
-To model reaction times using a drift diffusion model, we can use `td_ddm` (here, for speed, we are starting the optimization near optimal values):
+To model reaction times using a drift diffusion model, we can use `td_ddm` (here, for speed, we are starting the optimization near optimal values for this dataset):
 
 ```{r}
 ddm <- td_ddm(td_bc_single_ptpt, discount_function = 'exponential',
