+++ README.Rmd
@@ -121,7 +121,7 @@ print(ddm)
 
 Following [Peters & D'Esposito (2020)](https://doi.org/10.1371/journal.pcbi.1007615), we can apply a sigmoidal transform to the drift rate $\delta$ to improve model fit using the argument `drift_transform = "sigmoid"`:
 
-$$\delta' = v_\text{max} \left(\frac{2}{1 + \exp\{-\delta\}} - 1\right)$$
+$$\delta' = v_\text{max} \left(\frac{2}{1 + e^{-\delta}} - 1\right)$$
 
 ```{r}
 ddm_sig <- td_ddm(td_bc_single_ptpt, discount_function = 'exponential',
