### 2024-11-13 kinleyid/tempodisco#4 @marberts
No minimum version of R
Consider adding `Depends: R (>= x.y.z)` to `DESCRIPTION` to specify the minimum version of R needed to use the package. When installing with `devtools::install_github("kinleyid/tempodisco")` I got a warning related to this.
```
     NB: this package now depends on R (>= 3.5.0)
     WARNING: Added dependency on R >= 3.5.0 because serialized objects in
     serialize/load version 3 cannot be read in older versions of R.
     File(s) containing such objects:
       ‘tempodisco/data/adj_amt_sim.rda’
       ‘tempodisco/data/td_bc_single_ptpt.rda’
       ‘tempodisco/data/td_bc_study.rda’
       ‘tempodisco/data/td_ip_simulated_ptpt.rda’
```

If this is >= 4.4.0 then the utility function `%def%` can be replaced by the new null coalescing operator `%||%`.

Part of openjournals/joss-reviews#7472.
