### 2024-11-13 kinleyid/tempodisco#5 @marberts
Contributing guidelines
I could not find guidelines for contributing to the software, reporting issues/problems, or seeking support. Consider adding information in the readme or add `CONTRIBUTING.md` with this information.

Part of openjournals/joss-reviews#7472.
