### 2024-11-14 kinleyid/tempodisco#6 @marberts
Signature for `deviance()` and `logLik()` methods
Minor comment, but the signature for `deviance()` and `logLik()` methods is not the same as the generic (`function(mod)` instead of `function(object, ...)`).

Part of openjournals/joss-reviews#7472.
