### 2024-11-14 kinleyid/tempodisco#7 @marberts
Package documentation
Consider adding more/clarifying information in the readme about the problem your package is designed to solve and who the target audience is. I figured it out after going through the readme and looking at some of the references, but something a bit more explicit would be helpful. It may also be helpful to put the target audience in the paper.

You may also want to add a bit more to the `DESCRIPTION` along with some references to the literature. I've had initial submissions to CRAN rejected before because it wasn't clear where it fit in the literature/what methods were being implemented.

A few small suggestions to make the documentation a bit easier to use.
- [x] Have a package-level manual page (`package?tempodisco`) with the usual information. This was the first thing I tried after installing the package to get started.
- [x] Use `@family` to link related manual pages (e.g., have links to some of the methods for `td_bcnm` objects on the help page for `td_bcnm()`). You may also want to cross reference some of the related functions like `td_bclm()` and `td_bcnm()`.
- [x] Group the help pages for the pkgdown site instead of having them in one big list.
- [x] Add a url for the pkgdown site to the "About" sidebar on Github (I had to clone the repo and build it myself).
- [x] Consider adding an introductory "Getting started" vignette. This would also free up real estate on the readme for things like contributor guidelines.
- [x] Some parameter descriptions in the help pages end with a period, some don't.

Part of openjournals/joss-reviews#7472.
