### 2024-11-14 kinleyid/tempodisco#8 @marberts
Exisiting tools for temporal discounting
How does this package relate to existing tools? For example, https://github.com/miyamot0/discountingtools. There should probably be some mention in the paper of how this package contributes to existing tools for modelling temporal discounting.

Part of openjournals/joss-reviews#7472.
