### 2024-11-16 kinleyid/tempodisco#9 @marberts
CRAN checks
Running `R CMD check --as-cran` gives some notes that should probably be fixed before releasing the package.

>  Found the following (possibly) invalid URLs:
    URL: https://codecov.io/github/kinleyid/tempodisco (moved to https://app.codecov.io/github/kinleyid/tempodisco)
      From: README.md
      Status: 200
      Message: OK
    URL: https://cran.r-project.org/web/packages/RWiener/index.html
      From: inst/doc/drift-diffusion-models.html
      Status: 200
      Message: OK
      CRAN URL not in canonical form
    The canonical URL of the CRAN page for a package is
      https://CRAN.R-project.org/package=pkgname

  > The Title field should be in title case. Current version is:
  ‘Temporal discounting models’
  In title case that is:
  ‘Temporal Discounting Models’

  > Found the following URLs which should use \doi (with the DOI name only):
    File ‘kirby_consistency.Rd’:
      https://doi.org/10.1037//0096-3445.128.1.78
      https://doi.org/10.1007/s40614-016-0070-9
    File ‘kirby_score.Rd’:
      https://doi.org/10.1037//0096-3445.128.1.78
      https://doi.org/10.1007/s40614-016-0070-9
    File ‘predict.td_ddm.Rd’:
      https://doi.org/10.1016/j.jmp.2009.01.006
    File ‘td_ddm.Rd’:
      https://doi.org/10.1371/journal.pcbi.1007615
      https://doi.org/10.3758/s13423-018-1554-2
    File ‘wileyto_score.Rd’:
      https://doi.org/10.3758/BF03195548

You may also want to look at some of the suggestions from `goodpractice`.

Part of openjournals/joss-reviews#7472.
