### 2024-11-16 kinleyid/tempodisco#10 @marberts
Custom discount function without starting params
It is not clear what arguments are optional for `td_fn()`. For example, the default value for `par_starts` doesn't give a workable discount function, something that only shows up when trying to use the function.This argument isn't optional and there probably shouldn't be a default value.

``` r
library(tempodisco)

data("td_ip_simulated_ptpt")

exp_discount <- td_fn(
  name = "custom_exp",
  fn = function(data, p) exp(-p["k"] *data$del),
  par_lims = list(k = c(0, Inf))
)

td_ipm(td_ip_simulated_ptpt, discount_function = exp_discount)
#> Error in do.call(expand.grid, par_starts): second argument must be a list

exp_discount <- td_fn(
  name = "custom_exp",
  fn = function(data, p) exp(-p["k"] *data$del),
  par_starts = list(k = 0),
  par_lims = list(k = c(0, Inf))
)

td_ipm(td_ip_simulated_ptpt, discount_function = exp_discount)
#>
#> Temporal discounting indifference point model
#>
#> Discount function: custom_exp, with coefficients:
#>
#>           k
#> 0.001226789
#>
#> ED50: 565.009156750231
#> AUC: 0.0370079083492731
```

Omitting the `name` argument will similarly give an error when trying to print the model.

Part of openjournals/joss-reviews#7472.
