### 2024-11-16 cbd9111
More sensible error without param starts (#10)
