### 2024-11-16 #7472 @marberts
I'm done my initial review. I've opened several issues, the most important of which are kinleyid/tempodisco#7 (who is the target audience) and kinleyid/tempodisco#8. kinleyid/tempodisco#4 is the reason that installation instructions is unchecked and  kinleyid/tempodisco#5 is the reason that community guidelines is not checked. The issues and the PR are just suggestions to improve the code base a bit, especially if you plan to submit to CRAN.
