### 2024-11-18 kinleyid/tempodisco#12 @marberts
Input validation for kirby_score
`kibry_score()` doesn't validate `data`, which is particularly problematic because that argument isn't well documented. This can result is unhelpful errors.

``` r
library(tempodisco)

data("td_bc_single_ptpt")

td_bc_single_ptpt$imm_chosen <- as.character(td_bc_single_ptpt$imm_chosen)

kirby_score(td_bc_single_ptpt)
#> Error in !data$imm_chosen[(idx):(nrow(data) + 1)]: invalid argument type

td_bc_single_ptpt$imm_chosen <- NA

kirby_score(td_bc_single_ptpt)
#> Error in if (max_consistency < 0.75) {: missing value where TRUE/FALSE needed
```

Part of openjournals/joss-reviews#7472.
