### 2024-11-18 4070145
Response to reviewer comments up to pkgdown
