+++ paper/paper.bib
@@ -111,3 +111,37 @@
   doi={10.1371/journal.pcbi.1007615},
   publisher={Public Library of Science San Francisco, CA USA}
 }
+
+@article{odum2011delay,
+  title={Delay discounting: I'm ak, you're ak},
+  author={Odum, Amy L},
+  journal={Journal of the experimental analysis of behavior},
+  volume={96},
+  number={3},
+  pages={427--439},
+  year={2011},
+  doi={10.1901/jeab.2011.96-423},
+  publisher={Wiley Online Library}
+}
+
+@article{amlung2019delay,
+  title={Delay discounting as a transdiagnostic process in psychiatric disorders: A meta-analysis},
+  author={Amlung, Michael and Marsden, Emma and Holshausen, Katherine and Morris, Vanessa and Patel, Herry and Vedelago, Lana and Naish, Katherine R and Reed, Derek D and McCabe, Randi E},
+  journal={JAMA psychiatry},
+  volume={76},
+  number={11},
+  pages={1176--1186},
+  year={2019},
+  publisher={American Medical Association}
+}
+
+@article{story2014does,
+  title={Does temporal discounting explain unhealthy behavior? A systematic review and reinforcement learning perspective},
+  author={Story, Giles W and Vlaev, Ivo and Seymour, Ben and Darzi, Ara and Dolan, Raymond J},
+  journal={Frontiers in behavioral neuroscience},
+  volume={8},
+  pages={76},
+  year={2014},
+  doi={10.3389/fnbeh.2014.00076},
+  publisher={Frontiers Media SA}
+}
