+++ R/td_bcnm.R
@@ -13,6 +13,7 @@
 #' @param silent Boolean (true by default). The call to \code{optim()} occurs within a \code{try()} wrapper. The value of \code{silent} is passed along to \code{try()}.
 #' @param na.action Action to take when data contains \code{NA} values. Default is \code{na.omit}.
 #' @param ... Additional arguments to provide finer-grained control over the model configuration.
+#' @family nonlinear binary choice model functions
 #' @return An object of class \code{td_bcnm} with components \code{data} (containing the data used for fitting), \code{config} (containing the internal configuration of the model, including the \code{discount_function}), and \code{optim} (the output of \code{optim()}).
 #' @examples
 #' \dontrun{
