### 2024-11-18 kinleyid/tempodisco#4 @kinleyid
Added a minimum R version in 4070145e3506fc26c0339232fba41145422949db! I've decided to stick with 3.5.0 for now but am happy to up to 4.4.0 if needed.
