+++ README.md
@@ -9,7 +9,17 @@
 [![codecov](https://codecov.io/github/kinleyid/tempodisco/graph/badge.svg?token=CCQXS3SNGB)](https://codecov.io/github/kinleyid/tempodisco)
 <!-- badges: end -->
 
-`tempodisco` is an R package for working with delay discounting data.
+`tempodisco` is an R package for behavioural researchers working with
+delay discounting data (also known as temporal discounting intertemporal
+choice data). It is intended to simplify common tasks such as scoring
+responses (e.g. computing indifference points from an adjusting amounts
+procedure, computing the “area under the curve”, or computing $k$ values
+as in the Monetary Choice Questionnaire; [Kirby et al.,
+1999](https://doi.org/10.1037//0096-3445.128.1.78)), identifying
+poor-quality data (e.g. non-systematic responding and failed attention
+checks), modelling choice data using multiple discount functions
+(e.g. hyperbolic, exponential, etc.—see below), and modelling reaction
+times using drift diffusion models.
 
 ## Installation
 
@@ -48,15 +58,15 @@ indifference points. The function `td_ipm` can then be used to identify
 the best-fitting discount function (according to the Bayesian
 information criterion) from any subset of the following options:
 
-| Name                                                                                         | Functional form                                         |
-|----------------------------------------------------------------------------------------------|---------------------------------------------------------|
-| Exponential ([Samuelson, 1937](https://doi.org/10.2307/2967612))                             | $f(t; k) = e^{-k t}$                                    |
-| Scaled exponential (beta-delta; [Laibson, 1997](https://doi.org/10.1162/003355397555253))    | $f(t; k, w) = w e^{-k t}$                               |
-| Nonlinear-time exponential ([Ebert & Prelec, 2007](https://doi.org/10.1287/mnsc.1060.0671))  | $f(t; k, s) = e^{-k t^s}$                               |
-| Dual-systems exponential ([Ven den Bos & McClure, 2013](https://doi.org/10.1002/jeab.6))     | $f(t; k_1, k_2, w) = w e^{-k_1 t} + (1 - w) e^{-k_2 t}$ |
-| Inverse q-exponential ([Green & Myerson, 2004](https://doi.org/10.1037/0033-2909.130.5.769)) | $f(t; k, s) = \frac{1}{(1 + k t)^s}$                    |
-| Hyperbolic ([Mazur, 1987](https://doi.org/10.4324/9781315825502))                            | $f(t; k) = \frac{1}{1 + kt}$                            |
-| Nonlinear-time hyperbolic ([Rachlin, 2006](https://doi.org/10.1901/jeab.2006.85-05))         | $f(t; k, s) = \frac{1}{1 + k t^s}$                      |
+| Name | Functional form |
+|----|----|
+| Exponential ([Samuelson, 1937](https://doi.org/10.2307/2967612)) | $f(t; k) = e^{-k t}$ |
+| Scaled exponential (beta-delta; [Laibson, 1997](https://doi.org/10.1162/003355397555253)) | $f(t; k, w) = w e^{-k t}$ |
+| Nonlinear-time exponential ([Ebert & Prelec, 2007](https://doi.org/10.1287/mnsc.1060.0671)) | $f(t; k, s) = e^{-k t^s}$ |
[…]
