+++ pkgdown.yml
@@ -11,7 +11,7 @@ articles:
   modeling-binary-choice-data: modeling-binary-choice-data.html
   nonsystematic-discounting: nonsystematic-discounting.html
   visualizing-models: visualizing-models.html
-last_built: 2024-11-18T23:51Z
+last_built: 2024-11-19T00:02Z
 urls:
   reference: https://kinleyid.github.io/tempodisco/reference
   article: https://kinleyid.github.io/tempodisco/articles
