### 2024-11-19 kinleyid/tempodisco#5 @kinleyid
Added these guidelines to the readme in 4a78a75dbe87ab82442a86c6aa506b235786d65b!
