### 2024-11-19 b3ee8e6
Update function signatures
