+++ man/logLik.td_ipm.Rd
@@ -8,6 +8,8 @@
 }
 \arguments{
 \item{mod}{An object of class \code{td_ipm}}
+
+\item{...}{Additional arguments currently not used.}
 }
 \description{
 Compute log-likelihood for a temporal discounting indifference point model.
