+++ man/logLik.td_ddm.Rd
@@ -10,6 +10,8 @@
 \item{mod}{An object of class \code{td_bcnm}}
 
 \item{type}{Should probabilities /probability densities be computed for responses and RTs (\code{'resp_rt'}, default) or responses only (\code{'resp'})?}
+
+\item{...}{Additional arguments currently not used.}
 }
 \description{
 Compute log-likelihood for a temporal discounting drift diffusion model
