+++ man/logLik.td_bcnm.Rd
@@ -8,6 +8,8 @@
 }
 \arguments{
 \item{mod}{An object of class \code{td_bcnm}}
+
+\item{...}{Additional arguments currently not used.}
 }
 \description{
 Compute log-likelihood for a temporal discounting binary choice model.
