+++ man/deviance.td_ddm.Rd
@@ -8,6 +8,8 @@
 }
 \arguments{
 \item{mod}{An object of class \code{td_ddm}}
+
+\item{...}{Additional arguments currently not used.}
 }
 \description{
 Compute deviance for a temporal discounting drift diffusion model
