### 2024-11-19 kinleyid/tempodisco#7 @kinleyid
Just hoping to check my understanding of a few of these:
- I've set up a pkgdown site, which automatically organizes the help pages under docs/. Is this what the third point means?
- Would the idea be to replace the "Example usage" section of the readme with a link to the pkgdown page for the "Getting started" vignette?

Otherwise, the points I've checked off so far are addressed in b3ee8e60e8069ee5677cce297c792155590c245a!
