### 2024-11-19 kinleyid/tempodisco#6 @kinleyid
These function signatures are updated in b3ee8e60e8069ee5677cce297c792155590c245a!
