+++ pkgdown.yml
@@ -7,11 +7,12 @@ articles:
   comparing-models: comparing-models.html
   custom-discount-functions: custom-discount-functions.html
   drift-diffusion-models: drift-diffusion-models.html
+  getting-started: getting-started.html
   model-free-auc: model-free-auc.html
   modeling-binary-choice-data: modeling-binary-choice-data.html
   nonsystematic-discounting: nonsystematic-discounting.html
   visualizing-models: visualizing-models.html
-last_built: 2024-11-19T00:02Z
+last_built: 2024-11-19T00:16Z
 urls:
   reference: https://kinleyid.github.io/tempodisco/reference
   article: https://kinleyid.github.io/tempodisco/articles
