+++ paper/paper.bib
@@ -145,3 +145,45 @@
   doi={10.3389/fnbeh.2014.00076},
   publisher={Frontiers Media SA}
 }
+
+@manual{@discountingtools,
+  title = {discountingtools: Delay Discounting Tools},
+  author = {Shawn Gilroy},
+  year = {2024},
+  note = {R package version 0.0.3.2, commit 712ca3313a9eee44abb03c56c936d2141f90ce27},
+  url = {https://github.com/miyamot0/discountingtools},
+  doi={10.5281/zenodo.7754483}
+}
+
+@article{vincent2016hierarchical,
+  title={Hierarchical Bayesian estimation and hypothesis testing for delay discounting tasks},
+  author={Vincent, Benjamin T},
+  journal={Behavior research methods},
+  volume={48},
+  number={4},
+  pages={1608--1620},
+  year={2016},
+  publisher={Springer}
+}
+
+@article{kaplan2016automating,
+  title={Automating scoring of delay discounting for the 21-and 27-item monetary choice questionnaires},
+  author={Kaplan, Brent A and Amlung, Michael and Reed, Derek D and Jarmolowicz, David P and McKerchar, Todd L and Lemley, Shea M},
+  journal={The Behavior Analyst},
+  volume={39},
+  pages={293--304},
+  year={2016},
+  publisher={Springer}
+}
+
+@article{wulff2018modeling,
+  title={Modeling choices in delay discounting},
+  author={Wulff, Dirk U and van den Bos, Wouter},
+  journal={Psychological Science},
+  volume={29},
+  number={11},
+  pages={1890--1894},
+  year={2018},
+  doi={10.1177/0956797616664342},
+  publisher={Sage Publications Sage CA: Los Angeles, CA}
+}
