### 2024-11-19 kinleyid/tempodisco#8 @kinleyid
I've added a paragraph to the paper on related software tools in 8cc54d164b3f6feefb6d820be53ad056f772c9dd!
