### 2024-11-19 790b4c1
Updates based on output of gp()
