### 2024-11-19 13a6572
Add another software reference to paper.md
