+++ paper/paper.bib
@@ -187,3 +187,11 @@
   doi={10.1177/0956797616664342},
   publisher={Sage Publications Sage CA: Los Angeles, CA}
 }
+
+@manual{discAUC,
+  title = {discAUC: Linear and Non-Linear AUC for Discounting Data},
+  author = {Jonathan E. Friedel},
+  year = {2024},
+  note = {R package version 1.0.3},
+  url = {https://CRAN.R-project.org/package=discAUC},
+}
