+++ paper/paper.md
@@ -30,7 +30,7 @@ bibliography: paper.bib
 
 `tempodisco` is an R package for working with temporal discounting data aimed at behavioural researchers in psychology, economics, and related fields. It is designed to simplify the analysis steps of data cleaning, model fitting, and quantifying discounting. The package implements widely used methods such as computing indifference points from adjusting amount task [@frye2016measuring], testing for non-systematic discounting per the criteria of @johnson2008algorithm, scoring questionnaires according to the methods of @kirby1999heroin and @wileyto2004using, approximate Bayesian model selection using a range of discount functions [@franck2015accurate] for models of both choice-level data and indifference points, drift diffusion models of discounting [@peters2020drift], and model-agnostic measures of discounting such as area under the curve [@myerson2001area] and ED50 [@yoon2008turning]. Users can also produce simple diagnostic plots to examine model fits. Moreover, the package provides a framework within which users can specify custom discount functions.
 
-So far, there are relatively few software tools dedicated to working with temporal discounting data: a widely-used Excel-based spreadsheet tool [@kaplan2016automating] implements the @kirby1999heroin scoring method; a Matlab toolbox [@vincent2016hierarchical] implements hierarchical Bayesian modelling of choice-level data; and the R package `discountingtools` [@discountingtools] fits a range of discount functions to indifference point data. `tempodisco` provides several important functions not available in these tools, such as modeling choice-level data using a range of discount functions, fitting drift diffusion models, and implementing multiple "choice rules" to link subjective valuations to choice probabilities [@wulff2018modeling]. However, `tempodisco` currently fits only individual-level models (vs hierarchical models) and only models choices between immediate and delayed rewards (vs choices between two delayed rewards).
+So far, there are relatively few software tools dedicated to working with temporal discounting data: a widely-used Excel-based spreadsheet tool [@kaplan2016automating] implements the @kirby1999heroin scoring method; a Matlab toolbox [@vincent2016hierarchical] implements hierarchical Bayesian modelling of choice-level data; the R package `discountingtools` [@discountingtools] fits a range of discount functions to indifference point data; and the R package `discAUC` [@discAUC] computes the model-free area under the curve for indifference point data. `tempodisco` provides several important functions not available in these tools, such as modeling choice-level data using a range of discount functions, fitting drift diffusion models, and implementing multiple "choice rules" to link subjective valuations to choice probabilities [@wulff2018modeling]. However, `tempodisco` currently fits only individual-level models (vs hierarchical models) and only models choices between immediate and delayed rewards (vs choices between two delayed rewards).
 
 The package is based around a few basic objects such as `td_fn` for discount functions, `td_ipm` for models of indifference points, `td_bclm`/`td_bcnm` for linear/nonlinear models of binary choice data, and `td_ddm` for drift diffusion models. Objects containing models are designed to feel much like R's `lm` and `glm` objects, with methods for generic functions such as `coef`, `predict`,  `residuals`, and `A/BIC`. Thus, users familiar with basic R functionality should find it straightforward to use `tempodisco`. With the growing recognition that temporal discounting may be described by a range of functions beyond the standard hyperbolic [@franck2015accurate; @bailey2021problems] and that model fit should be assessed by measures other than $R^2$ [@gelinor], `tempodisco` will allow researchers to easily implement these best practices and will continue to evolve as best practices in temporal discounting research do.
 
