### 2024-11-19 3878eea
Bug fixes from misunderstanding vapply()
