+++ pkgdown.yml
@@ -12,7 +12,7 @@ articles:
   modeling-binary-choice-data: modeling-binary-choice-data.html
   nonsystematic-discounting: nonsystematic-discounting.html
   visualizing-models: visualizing-models.html
-last_built: 2024-11-19T02:01Z
+last_built: 2024-11-19T03:25Z
 urls:
   reference: https://kinleyid.github.io/tempodisco/reference
   article: https://kinleyid.github.io/tempodisco/articles
