### 2024-11-19 c30c433
Added sections for pkgdown site (#7)
