### 2024-11-19 kinleyid/tempodisco#7 @marberts
I opened PR #14 giving an example of what I mean for the third point. I probably didn't group the functions entirely right.

For the fifth point, I think it would be better to have a getting started vignette (called `tempodisco.Rmd` so `pkgdown` gives it a special spot on the navbar) with the content of the readme. You can then have a really short example in the readme, or just refer to the vingette. That way the readme stays nice and short. But that's also my preference; feel free to do it differently.
