### 2024-11-19 bd617ee
Add util to validate data
