### 2024-11-19 cfaccb3
Use new validation function instead of require_columns
