+++ R/td_bclm.R
@@ -39,10 +39,6 @@ td_bclm <- function(data,
   # Validate data
   validate_td_data(data,
                    required_columns = c('val_imm', 'val_del', 'del', 'imm_chosen'))
-  req_cols <- c('val_imm', 'val_del', 'del', 'imm_chosen')
-  require_columns(data, req_cols)
-  data <- data[req_cols]
-  data$imm_chosen <- as.logical(data$imm_chosen)
   attention_checks(data, warn = T)
   invariance_checks(data, warn = T)
   if (length(grep('\\.B', names(data))) > 1) {
