### 2024-11-19 831aada
Debug new data validation function
