### 2024-11-19 50eed5a
Update adj_amt_indiffs
