### 2024-11-19 0502216
Reorganize internal-utils
Switch to using get_candidate_discount_functions instead of validate_discount_function
