### 2024-11-19 6a7757d
Remove now-unused discount function validator
