+++ R/td_ipm.R
@@ -61,24 +61,6 @@ td_ipm <- function(
   validate_td_data(data, required_columns = c('indiff', 'del'))
   data <- na.action(data)
   
-  # # Valid discount function name
-  # validate_discount_function(discount_function)
-  # 
-  # # Get a list of candidate td_fn objects to test
-  # if (is.character(discount_function)) {
-  #   # Vector of characters
-  #   cand_fns <- list()
-  #   for (fn_name in discount_function) {
-  #     cand_fns <- c(cand_fns, list(td_fn(fn_name)))
-  #   }
-  # } else {
-  #   if (is(discount_function, 'td_fn')) {
-  #     cand_fns <- list(discount_function) # List containing only one td_fn object
-  #   } else {
-  #     cand_fns <- discount_function # assumed to already be a list of td_fn objects, ensured by validation above
-  #   }
-  # }
-  
   cand_fns <- get_candidate_discount_functions(discount_function)
   
   # Run optimization on each candidate discount function
