### 2024-11-19 kinleyid/tempodisco#10 @kinleyid
`td_fn` now has more extensive input validation!
