### 2024-11-19 kinleyid/tempodisco#12 @kinleyid
I've created a data validation function (`validate_td_data()`) that's now used in many functions including `kirby_score()`
