### 2024-11-19 d5f7ffe
Consistent use of periods in documentation
