+++ R/generics.R
@@ -76,13 +76,13 @@ predict_indiffs <- function(object, newdata) {
 
 #' Model Predictions
 #'
-#' Generate predictions from a temporal discounting binary choice model
+#' Generate predictions from a temporal discounting binary choice model.
 #' @param object A temporal discounting binary choice model. See \code{\link{td_bcnm}}.
 #' @param newdata Optionally, a data frame to use for prediction. If omitted, the data used to fit the model will be used for prediction.
 #' @param type The type of prediction required. As in predict.glm, \code{"link"} (default) and \code{"response"} give predictions on the scales of the linear predictors and response variable, respectively. \code{"indiff"} gives predicted indifference points. For predicting indifference points, \code{newdata} needs only a \code{del} column.
 #' @param ... Additional arguments currently not used.
 #' @family nonlinear binary choice model functions
-#' @return A vector of predictions
+#' @return A vector of predictions.
 #' @examples
 #' \dontrun{
 #' data("td_bc_single_ptpt")
@@ -124,13 +124,13 @@ predict.td_bcnm <- function(object, newdata = NULL, type = c('link', 'response',
 
 #' Model Predictions
 #'
-#' Generate predictions from a temporal discounting binary choice linear model
+#' Generate predictions from a temporal discounting binary choice linear model.
 #' @param object A temporal discounting binary choice linear model. See \code{td_bclm}.
 #' @param newdata Optionally, a data frame to use for prediction. If omitted, the data used to fit the model will be used for prediction.
 #' @param type The type of prediction required. For \code{'indiff'} (default) gives predicted indifference points. In this case, \code{newdata} needs only a \code{del} column. For all other values (e.g. \code{"link"}, \code{"response"}), this function is just a wrapper to \code{predict.glm()}.
-#' @param ... Additional arguments passed to predict.glm if type != \code{'indiff'}
+#' @param ... Additional arguments passed to predict.glm if type != \code{'indiff'}.
 #' @family linear binary choice model functions
-#' @return A vector of predictions
+#' @return A vector of predictions.
 #' @examples
 #' \dontrun{
 #' data("td_bc_single_ptpt")
@@ -163,10 +163,10 @@ predict.td_bclm <- function(object, newdata = NULL, type = c('indiff', 'link', '
 #' Generate predictions from a temporal discounting indifference point model
 #' @param object A temporal discounting indifference point model. See \code{td_ipm}.
 #' @param newdata A data frame to use for prediction. If omitted, the data used to fit the model will be used for prediction.
-#' @param type Type of prediction, either \code{'indiff'} (indifference points) or \code{'response'} (whether the participants would is predicted to choose the immediate (1) or delayed reward (0))
+#' @param type Type of prediction, either \code{'indiff'} (indifference points) or \code{'response'} (whether the participants would is predicted to choose the immediate (1) or delayed reward (0)).
 #' @param ... Additional arguments currently not used.
 #' @family indifference point model functions
-#' @return A vector of predictions
+#' @return A vector of predictions.
 #' @examples
 #' \dontrun{
 #' data("td_ip_simulated_ptpt")
@@ -204,13 +204,13 @@ predict.td_ipm <- function(object, newdata = NULL, type = c('indiff', 'response'
 
 #' Model Predictions
 #'
-#' Generate predictions from a temporal discounting drift diffusion model
+#' Generate predictions from a temporal discounting drift diffusion model.
 #' @param object A temporal discounting drift diffusion model. See \code{\link{td_ddm}}.
[…]
