### 2024-11-20 c66b1b9
Update vignettes and AUC
