+++ tests/testthat/test-td_bclm.R
@@ -17,8 +17,7 @@ while (model_idx <= length(models)) {
   test_that('generics', {
     expect_output(print(mod))
     expect_in(class(ED50(mod)), c('numeric', 'character')) # Output might be "none", e.g. for model-free
-    expect_in(class(AUC(mod, verbose = F)), c('numeric', 'character'))
-    expect_output(AUC(mod, verbose = T))
+    expect_in(class(AUC(mod)), c('numeric', 'character'))
     expect_type(BIC(mod), 'double')
     expect_type(AIC(mod), 'double')
     expect_type(AIC(mod, k = 5), 'double')
