+++ R/tempodisco-package.R
@@ -1,5 +1,5 @@
 
-#' @importFrom stats optim predict qlogis plogis residuals integrate coef BIC AIC glm binomial fitted approx predict.glm na.omit median filter
+#' @importFrom stats optim predict qlogis plogis residuals integrate coef BIC AIC glm binomial fitted approx approxfun predict.glm na.omit median filter
 #' @importFrom graphics lines points title axis
 #' @importFrom methods is
 #' @importFrom grDevices rgb
