+++ pkgdown.yml
@@ -4,15 +4,15 @@ pkgdown_sha: ~
 articles:
   adjusting-amounts: adjusting-amounts.html
   analyzing-study: analyzing-study.html
+  area-under-curve: area-under-curve.html
   comparing-models: comparing-models.html
   custom-discount-functions: custom-discount-functions.html
   drift-diffusion-models: drift-diffusion-models.html
   getting-started: getting-started.html
-  model-free-auc: model-free-auc.html
   modeling-binary-choice-data: modeling-binary-choice-data.html
   nonsystematic-discounting: nonsystematic-discounting.html
   visualizing-models: visualizing-models.html
-last_built: 2024-11-19T23:33Z
+last_built: 2024-11-20T05:27Z
 urls:
   reference: https://kinleyid.github.io/tempodisco/reference
   article: https://kinleyid.github.io/tempodisco/articles
