### 2024-11-20 70fefe0
getting-started vignette and website customizations
