### 2024-11-20 0b44e97
Nicer plot colours
