### 2024-11-20 1a1d3b3
Shorten readme
