+++ README.md
@@ -6,341 +6,85 @@
 <!-- badges: start -->
 
 [![R-CMD-check](https://github.com/kinleyid/tempodisco/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/kinleyid/tempodisco/actions/workflows/R-CMD-check.yaml)
-[![codecov](https://codecov.io/github/kinleyid/tempodisco/graph/badge.svg?token=CCQXS3SNGB)](https://codecov.io/github/kinleyid/tempodisco)
+[![codecov](https://codecov.io/github/kinleyid/tempodisco/graph/badge.svg?token=CCQXS3SNGB)](https://app.codecov.io/github/kinleyid/tempodisco)
 <!-- badges: end -->
 
 `tempodisco` is an R package for behavioural researchers working with
 delay discounting data (also known as temporal discounting intertemporal
-choice data). It is intended to simplify common tasks such as scoring
-responses (e.g. computing indifference points from an adjusting amounts
-procedure, computing the “area under the curve”, or computing $k$ values
-as in the Monetary Choice Questionnaire; [Kirby et al.,
+choice data). It implements common tasks such as scoring responses
+(e.g. computing indifference points from an adjusting amounts procedure,
+computing the “area under the curve”, or computing $k$ values as in the
+Monetary Choice Questionnaire; [Frye et al.,
+2016](https://doi.org/10.3791/53584); [Myerson et al.,
+2001](https://doi.org/10.1901/jeab.2001.76-235); [Kirby et al.,
 1999](https://doi.org/10.1037//0096-3445.128.1.78)), identifying
-poor-quality data (e.g. non-systematic responding and failed attention
-checks), modelling choice data using multiple discount functions
-(e.g. hyperbolic, exponential, etc.—see below), and modelling reaction
-times using drift diffusion models.
+poor-quality data (e.g. failed attention checks and non-systematic
+responding; [Johnson & Bickel,
+2008](https://doi.org/10.1037/1064-1297.16.3.264)), modelling choice
+data using multiple discount functions (e.g. hyperbolic, exponential,
+etc.—see below; [Franck et al.,
+2015](https://doi.org/10.1002/jeab.128)), and modelling reaction times
+using drift diffusion models ([Peters & D’Esposito,
+2020](https://doi.org/10.1371/journal.pcbi.1007615)).
 
 ## Installation
 
-You can install tempodisco from [GitHub](https://github.com/) with:
+You can install tempodisco from GitHub with:
 
 ``` r
 # install.packages("devtools")
 devtools::install_github("kinleyid/tempodisco")
 ```
 
-## Example usage
+## Getting started
+
+See [getting
+started](https://kinleyid.github.io/tempodisco/docs/articles/tempodisco.html)
+for example usage.
+
+## Overview
+
+A good practice in delay discounting research is to not assume that the
+same discount function describes every individual ([Franck et al.,
+2015](https://doi.org/10.1002/jeab.128)). `tempodisco` implements the
+following discount functions and can automatically select the best one
+for a given individual according to the Bayesian information criterion
[…]
