### 2024-11-20 0e1e880
Test using \doi{} instead of \href{}{}
