### 2024-11-20 cd297a4
Use \doi{} instead of \href{}{}
