+++ man/predict.td_ddm.Rd
@@ -27,7 +27,7 @@ A vector of predictions.
 Generate predictions from a temporal discounting drift diffusion model.
 }
 \note{
-When \code{type = 'rt'}, expected RTs are computed irrespective of which reward was selected, per equation 5 in \href{https://doi.org/10.1016/j.jmp.2009.01.006}{Grasman, Wagenmakers, & van der Maas (2009)}.
+When \code{type = 'rt'}, expected RTs are computed irrespective of which reward was selected, per equation 5 in Grasman, Wagenmakers, & van der Maas (2009, \doi{10.1016/j.jmp.2009.01.006}).
 }
 \examples{
 \dontrun{

[…]
