+++ pkgdown.yml
@@ -12,7 +12,7 @@ articles:
   nonsystematic-discounting: nonsystematic-discounting.html
   tempodisco: tempodisco.html
   visualizing-models: visualizing-models.html
-last_built: 2024-11-20T15:03Z
+last_built: 2024-11-20T15:12Z
 urls:
   reference: https://kinleyid.github.io/tempodisco/reference
   article: https://kinleyid.github.io/tempodisco/articles
