### 2024-11-20 kinleyid/tempodisco#9 @kinleyid
`R CMD check --as-cran` runs without warnings as of 4aaa247e9bb7f23d0f568d033fc5ba00dd14078b!
