### 2024-11-20 00e1eae
Add power discount function
