+++ R/td_fn.R
@@ -7,7 +7,7 @@
 #' @param fn Function that takes a data.frame and a vector of named parameters and returns a vector of values between 0 and 1.
 #' @param par_starts A named list of vectors, each specifying possible starting values for a parameter to try when running optimization.
 #' @param par_lims A named list of vectors, each specifying the bounds to impose of a parameters. Any parameter for which bounds are unspecified are assumed to be unbounded.
-#' @param ED50 A function which, given a named vector of parameters \code{p} and optionally a value of \code{del_val}, computes the ED50. If there is no closed-form solution, this should return the string "non-analytic". If the ED50 is not well-defined, this should return the string "none".
+#' @param ED50 A function which, given a named vector of parameters \code{p} and optionally a value of \code{del_val}, computes the ED50. If there is no closed-form solution, this should return the string "non-analytic". If the ED50 is not well-defined, this should return the string "none". As a shortcut for these latter 2 cases, the strings "non-analytic" and "none" can be directly supplied as arguments.
 #' @param par_chk Optionally, this is a function that checks the parameters to ensure they meet some criteria. E.g., for the dual-systems-exponential discount function, we require k1 < k2.
 #' @return An object of class `td_fn`.
 #' @examples 
@@ -26,14 +26,15 @@
 #' }
 #' @export
 td_fn <- function(predefined = c('hyperbolic',
-                                  'exponential',
-                                  'inverse-q-exponential',
-                                  'nonlinear-time-hyperbolic',
-                                  'scaled-exponential',
-                                  'dual-systems-exponential',
-                                  'nonlinear-time-exponential',
-                                  'model-free',
-                                  'constant'),
+                                 'exponential',
+                                 'power',
+                                 'inverse-q-exponential',
+                                 'nonlinear-time-hyperbolic',
+                                 'scaled-exponential',
+                                 'dual-systems-exponential',
+                                 'nonlinear-time-exponential',
+                                 'model-free',
+                                 'constant'),
                   name = NULL,
                   fn = NULL,
                   par_starts = NULL,
@@ -56,6 +57,9 @@ td_fn <- function(predefined = c('hyperbolic',
       out$fn <- fn
     }
     
+    if (is.null(par_starts)) {
+      stop('par_starts must ')
+    }
     if (is.null(names(par_starts))) {
       stop('par_starts must be a named list')
     } else {
@@ -81,8 +85,18 @@ td_fn <- function(predefined = c('hyperbolic',
     
     if (is.null(ED50)) {
       out$ED50 <- function(...) 'non-analytic'
-    } else {
+    } else if (is.function(ED50)) {
       out$ED50 <- ED50
+    } else if (is.character(ED50)) {
+      if (ED50 == 'non-analytic') {
+        out$ED50 <- function(...) 'non-analytic'
+      } else if (ED50 == 'none') {
+        out$ED50 <- function(...) 'none'
+      } else {
+        stop('Only "non-analytic" and "none" are allowed')
+      }
+    } else {
+      stop('ED50 must be a function, "none", or "non-analytic"')
     }
     
     if (!is.null(par_chk)) {
@@ -92,137 +106,156 @@ td_fn <- function(predefined = c('hyperbolic',
   } else {
     
     name <- match.arg(predefined)
-    out$name <- name
-    out$fn <- switch (name,
-                      'constant' = function(data, p) rep(p['c'], length(data$del)),
-                      'hyperbolic' = function(data, p) 1 / (1 + p['k']*data$del),
-                      'exponential' = function(data, p) exp(-p['k']*data$del),
-                      'inverse-q-exponential' = function(data, p) 1 / (1 + p['k']*data$del)**p['s'],
[…]
