### 2024-11-20 a8a1bc0
Fix model-free discount function
