### 2024-11-20 a72e6b3
Add power discount function to td_bclm
