+++ man/td_bclm.Rd
@@ -7,7 +7,8 @@
 td_bclm(
   data,
   model = c("all", "hyperbolic.1", "hyperbolic.2", "exponential.1", "exponential.2",
-    "scaled-exponential", "nonlinear-time-hyperbolic", "nonlinear-time-exponential"),
+    "scaled-exponential", "nonlinear-time-hyperbolic", "power",
+    "nonlinear-time-exponential"),
   ...
 )
 }
