### 2024-11-22 6d12217
Make "model-free" discount function more robust
Previously, the correspondence between delays and indifference points was encoded in the names of the parameter vector. This wasn't great because it required storing floating point data in strings. Now the correspondence is established in an enclosing environment when an init() function is called.
