+++ R/td_fn.R
@@ -7,6 +7,7 @@
 #' @param fn Function that takes a data.frame and a vector of named parameters and returns a vector of values between 0 and 1.
 #' @param par_starts A named list of vectors, each specifying possible starting values for a parameter to try when running optimization.
 #' @param par_lims A named list of vectors, each specifying the bounds to impose of a parameters. Any parameter for which bounds are unspecified are assumed to be unbounded.
+#' @param init A function to initialize the td_fn object. It should take 2 arguments: "self" (the td_fn object being initialized) and "data" (the data used for initialization).
 #' @param ED50 A function which, given a named vector of parameters \code{p} and optionally a value of \code{del_val}, computes the ED50. If there is no closed-form solution, this should return the string "non-analytic". If the ED50 is not well-defined, this should return the string "none". As a shortcut for these latter 2 cases, the strings "non-analytic" and "none" can be directly supplied as arguments.
 #' @param par_chk Optionally, this is a function that checks the parameters to ensure they meet some criteria. E.g., for the dual-systems-exponential discount function, we require k1 < k2.
 #' @return An object of class `td_fn`.
@@ -20,7 +21,7 @@
 #'   fn = function(data, p) (1 - p['b'])*exp(-p['k']*data$del) + p['b'],
 #'   par_starts = list(k = c(0.001, 0.1), b = c(0.001, 0.1)),
 #'   par_lims = list(k = c(0, Inf), b = c(0, 1)),
-#'   ED50 = function(...) 'non-analytic'
+#'   ED50 = 'non-analytic'
 #' )
 #' mod <- td_bcnm(td_bc_single_ptpt, discount_function = custom_discount_function, fit_err_rate = T)
 #' }
@@ -35,10 +36,11 @@ td_fn <- function(predefined = c('hyperbolic',
                                  'nonlinear-time-exponential',
                                  'model-free',
                                  'constant'),
-                  name = NULL,
+                  name = 'unnamed',
                   fn = NULL,
                   par_starts = NULL,
                   par_lims = NULL,
+                  init = NULL,
                   ED50 = NULL,
                   par_chk = NULL) {
   
@@ -47,46 +49,67 @@ td_fn <- function(predefined = c('hyperbolic',
   
   if (missing(predefined)) {
     
-    if (missing(name) | missing(fn) | missing(par_starts)) {
-      stop('To create a custom discount funciton, "name", "fn", and "par_starts" must all be provided to td_fn')
-    }
+    stopifnot(
+      is.character(name),
+      length(name) == 1
+    )
     out$name <- name
-    if (!all(names(formals(fn)) == c('data', 'p'))) {
-      stop('fn must take 2 arguments: data (a dataframe) and p (a vector of named parameters)')
-    } else {
-      out$fn <- fn
-    }
     
-    if (is.function(par_starts)) {
-      out$par_starts <- par_starts
-    } else {
-      if (is.null(names(par_starts))) {
-        stop('par_starts must be a named list')
+    if (missing(fn)) {
+      if (missing(init)) {
+        stop('fn must be supplied if it will not be created by init')
       } else {
-        out$par_starts <- as.list(par_starts)
+        fn <- function(data, p) NA
       }
+    } else {
+      stopifnot(is.function(fn))
+      if (!all(names(formals(fn)) == c('data', 'p')))
+        stop('fn must take 2 arguments: data (a dataframe) and p (a vector of named parameters)')
     }
+    out$fn <- fn
     
-    if (is.null(par_lims)) {
+    if (missing(par_starts)) {
+      if (missing(init)) {
+        stop('par_starts must be supplied if it will not be created by init')
+      } else {
+        par_starts <- list(placeholder = NA)
+      }
+    } else {
+      stopifnot(
+        is.list(par_starts),
+        !is.null(names(par_starts)),
[…]
