+++ tests/testthat/test-td_fn.R
@@ -38,14 +38,14 @@ test_that('errors', {
                      fn = function(data, p) {p['x']},
                      par_starts = c(x = 0.5), #!
                      par_lims = list(c(0, 1))))
-  expect_no_error(td_fn(name = 'custom',
-                        fn = function(data, p) {p['x']},
-                        par_starts = list(x = 0.5),
-                        par_lims = list(x = c(0, 1)),
-                        ED50 = 'random')) #!
-  expect_no_error(td_fn(name = 'custom',
-                        fn = function(data, p) {p['x']},
-                        par_starts = list(x = 0.5),
-                        par_lims = list(x = c(0, 1)),
-                        ED50 = 1)) #!
+  expect_error(td_fn(name = 'custom',
+              fn = function(data, p) {p['x']},
+              par_starts = list(x = 0.5),
+              par_lims = list(x = c(0, 1)),
+              ED50 = 'random')) #!
+  expect_error(td_fn(name = 'custom',
+                     fn = function(data, p) {p['x']},
+                     par_starts = list(x = 0.5),
+                     par_lims = list(x = c(0, 1)),
+                     ED50 = 1)) #!
 })
