+++ tests/testthat/test-methods.R
@@ -5,9 +5,10 @@ test_that('AUC on dataframes', {
   expect_no_error(AUC(td_ip_simulated_ptpt))
   expect_no_error(AUC(td_ip_simulated_ptpt, del_transform = 'log'))
   expect_no_error(AUC(td_ip_simulated_ptpt, del_transform = 'ord'))
+  expect_error(AUC(td_ip_simulated_ptpt,
+                   max_del = max(td_ip_simulated_ptpt) + 1))
   td_ip_simulated_ptpt$indiff <- NULL
   expect_error(AUC(td_ip_simulated_ptpt))
-  expect_error(AUC(td_ip_simulated_ptpt, max_del = max(td_ip_simulated_ptpt) + 1))
 })
 
 test_that('model-based AUC', {
