### 2024-11-23 00eeba7
Downgrade from macos-latest to macos-13
To get around the issue discussed here: https://github.com/r-lib/actions/issues/948
