### 2024-12-07 6d7fef8
Deal with NA values by simply omitting
