+++ R/td_bcnm.R
@@ -11,7 +11,6 @@
 #' @param eps_par_starts A vector of starting values to try for the "eps" parameter (which controls the error rate) during optimization. Ignored if `fit_err_rate = FALSE`.
 #' @param optim_args Additional arguments to pass to \code{optim()}. Default is \code{list(silent = T)}.
 #' @param silent Boolean (true by default). The call to \code{optim()} occurs within a \code{try()} wrapper. The value of \code{silent} is passed along to \code{try()}.
-#' @param na.action Action to take when data contains \code{NA} values. Default is \code{na.omit}.
 #' @param ... Additional arguments to provide finer-grained control over the model configuration.
 #' @family nonlinear binary choice model functions
 #' @return An object of class \code{td_bcnm} with components \code{data} (containing the data used for fitting), \code{config} (containing the internal configuration of the model, including the \code{discount_function}), and \code{optim} (the output of \code{optim()}).
@@ -40,7 +39,6 @@ td_bcnm <- function(
     eps_par_starts = c(0.01, 0.25),
     silent = T,
     optim_args = list(),
-    na.action = na.omit,
     ...) {
   
   # From a user's POV, it's easier to specify `choice_rule` and `fixed_ends`
@@ -86,10 +84,11 @@ td_bcnm <- function(
     }
   }
   
-  # Required data columns
-  validate_td_data(data,
-                   required_columns = c('val_imm', 'val_del', 'del', 'imm_chosen'))
-  data <- na.action(data)
+  # Data validation
+  data <- validate_td_data(
+    data,
+    required_columns = c('val_imm', 'val_del', 'del', 'imm_chosen')
+  )
   
   # Ensure imm_chosen is logical
   data$imm_chosen <- as.logical(data$imm_chosen)
@@ -172,7 +171,7 @@ td_bcnm <- function(
         )
       )
     }
-    # Run optimization
+    # Run optimizationn
     optimized <- run_optimization(nll_fn,
                                   par_starts,
                                   par_lims,

[…]
