### 2025-01-16 kinleyid/tempodisco#18 @bkrayfield
State of the field
The JOSS paper is missing the following section.

State of the field: Do the authors describe how this software compares to other commonly-used packages?

Even if other options do not exist, that is not stated in the paper.

Part of https://github.com/openjournals/joss-reviews/issues/7472
