### 2025-01-16 #7472 @bkrayfield
Hello Everyone, since I am coming behind the other reviewer, I agree with the comments that were made by @marberts. I have also noted in [kinleyid/tempodisco#18](https://github.com/kinleyid/tempodisco/issues/18) there is a missing section to the short paper.
