+++ paper/paper.bib
@@ -146,7 +146,7 @@
   publisher={Frontiers Media SA}
 }
 
-@manual{@discountingtools,
+@article{@discountingtools,
   title = {discountingtools: Delay Discounting Tools},
   author = {Shawn Gilroy},
   year = {2024},
