### 2025-01-17 kinleyid/tempodisco#18 @kinleyid
Apologies, I added this section a while ago but forgot to re-generate the PDF in the review issue. The most recent version of the paper, which describes other software packages in the second paragraph of the "Statement of need" section, is available here: https://github.com/openjournals/joss-reviews/issues/7472#issuecomment-2598363313. I'm having trouble getting one of the references to render properly, but the paper itself is still available in the meantime.
