### 2025-02-07 8243ff2
Adding a few new references
