+++ paper/paper.bib
@@ -195,3 +195,23 @@
   note = {R package version 1.0.3},
   url = {https://CRAN.R-project.org/package=discAUC},
 }
+
+@article{kinley2025probabilistic,
+  title = {Probabilistic models of delay discounting: “Fixed-endpoint” psychometric curves improve plausibility and performance},
+  author = {Isaac Kinley and Joseph Oluwasola and Suzanna Becker},
+  volume = {124},
+  pages = {102902},
+  year = {2025},
+  issn = {0022-2496},
+  doi = {https://doi.org/10.1016/j.jmp.2025.102902}
+}
+
+@article{green1997rate,
+  title={Rate of temporal discounting decreases with amount of reward},
+  author={Green, Leonard and Myerson, Joel and McFadden, Edward},
+  journal={Memory \& cognition},
+  volume={25},
+  pages={715--723},
+  year={1997},
+  publisher={Springer}
+}
