+++ paper/paper.md
@@ -28,10 +28,10 @@ bibliography: paper.bib
 
 # Statement of need
 
-`tempodisco` is an R package for working with temporal discounting data aimed at behavioural researchers in psychology, economics, and related fields. It is designed to simplify the analysis steps of data cleaning, model fitting, and quantifying discounting. The package implements widely used methods such as computing indifference points from adjusting amount task [@frye2016measuring], testing for non-systematic discounting per the criteria of @johnson2008algorithm, scoring questionnaires according to the methods of @kirby1999heroin and @wileyto2004using, approximate Bayesian model selection using a range of discount functions [@franck2015accurate] for models of both choice-level data and indifference points, drift diffusion models of discounting [@peters2020drift], and model-agnostic measures of discounting such as area under the curve [@myerson2001area] and ED50 [@yoon2008turning]. Users can also produce simple diagnostic plots to examine model fits. Moreover, the package provides a framework within which users can specify custom discount functions.
+`tempodisco` is an R package for working with temporal discounting data aimed at behavioural researchers in psychology, economics, and related fields. It is designed to simplify the analysis steps of data cleaning, model fitting, and quantifying discounting. The package implements widely used methods such as computing indifference points from adjusting amount task [@frye2016measuring], testing for non-systematic discounting per the criteria of @johnson2008algorithm, scoring questionnaires according to the methods of @kirby1999heroin and @wileyto2004using, and computing model-agnostic measures of discounting such as area under the curve [@myerson2001area] and ED50 [@yoon2008turning]. The package can also perform approximate Bayesian model selection using a range of discount functions [@franck2015accurate] for models of both choice-level data and indifference points, and implements more recently proposed methods such as drift diffusion models of discounting [@peters2020drift] and "fixed-endpoint" psychometric curves [@kinley2025probabilistic]. Users can also produce simple diagnostic plots to examine model fits. Moreover, the package provides a framework within which users can specify custom discount functions (e.g., discount functions to account for the "magnitude effect"; @green1997rate).
 
-So far, there are relatively few software tools dedicated to working with temporal discounting data: a widely-used Excel-based spreadsheet tool [@kaplan2016automating] implements the @kirby1999heroin scoring method; a Matlab toolbox [@vincent2016hierarchical] implements hierarchical Bayesian modelling of choice-level data; the R package `discountingtools` [@discountingtools] fits a range of discount functions to indifference point data; and the R package `discAUC` [@discAUC] computes the model-free area under the curve for indifference point data. `tempodisco` provides several important functions not available in these tools, such as modeling choice-level data using a range of discount functions, fitting drift diffusion models, and implementing multiple "choice rules" to link subjective valuations to choice probabilities [@wulff2018modeling]. However, `tempodisco` currently fits only individual-level models (vs hierarchical models) and only models choices between immediate and delayed rewards (vs choices between two delayed rewards).
[…]
