+++ paper/paper.bib
@@ -198,6 +198,7 @@
 
 @article{kinley2025probabilistic,
   title = {Probabilistic models of delay discounting: “Fixed-endpoint” psychometric curves improve plausibility and performance},
+  journal = {Journal of Mathematical Psychology},
   author = {Isaac Kinley and Joseph Oluwasola and Suzanna Becker},
   volume = {124},
   pages = {102902},
