### 2025-02-07 #7472 @kinleyid
Just adding a few new references to the paper. @bkrayfield did you get a chance to look at updated paper with "state of the field" information?
