+++ paper/paper.md
@@ -32,6 +32,6 @@ bibliography: paper.bib
 
 So far, there are relatively few software tools dedicated to working with temporal discounting data: a widely-used Excel-based spreadsheet tool [@kaplan2016automating] implements the @kirby1999heroin scoring method; a Matlab toolbox [@vincent2016hierarchical] implements hierarchical Bayesian modelling of choice-level data; the R package `discountingtools` [@discountingtools] fits a range of discount functions to indifference point data; and the R package `discAUC` [@discAUC] computes the model-free area under the curve for indifference point data. `tempodisco` provides several important functions not available in these tools, such as modeling choice-level data using a range of discount functions, fitting drift diffusion models, and implementing multiple "choice rules" to link subjective valuations to choice probabilities [@wulff2018modeling]. However, `tempodisco` currently fits only individual-level models (vs hierarchical models), only models choices between immediate and delayed rewards (vs choices between two delayed rewards), and only performs maximum-likelihood (vs Bayesian) parameter estimation.
 
-The package is based around a few basic objects such as `td_fn` for discount functions, `td_ipm` for models of indifference points, `td_bclm`/`td_bcnm` for linear/nonlinear models of binary choice data, and `td_ddm` for drift diffusion models. Objects containing models are designed to feel much like R's `lm` and `glm` objects, with methods for generic functions such as `coef`, `predict`,  `residuals`, and `A/BIC`. Thus, users familiar with basic R functionality should find it straightforward to use `tempodisco`. With the growing recognition that temporal discounting may be described by a range of functions beyond the standard hyperbolic [@franck2015accurate; @bailey2021problems] and that model fit should be assessed by measures other than $R^2$ [@gelinor], `tempodisco` will allow researchers to easily implement these best practices and will continue to evolve along with best practices in temporal discounting research.
+The package is based around a few basic objects such as `td_fn` for discount functions, `td_ipm` for models of indifference points, `td_bclm`/`td_bcnm` for linear/nonlinear models of binary choice data, and `td_ddm` for drift diffusion models. Objects containing models are designed to feel much like R's `lm` and `glm` objects, with methods for generic functions such as `coef`, `predict`,  `residuals`, and `A/BIC`. Thus, users familiar with basic R functionality should find it straightforward to use `tempodisco`. For getting started, the package comes with an example dataset (described in @kinley2024goal) pre-formatted to `tempodisco`'s specifications. With the growing recognition that temporal discounting may be described by a range of functions beyond the standard hyperbolic [@franck2015accurate; @bailey2021problems] and that model fit should be assessed by measures other than $R^2$ [@gelinor], `tempodisco` will allow researchers to easily implement these best practices and will continue to evolve along with best practices in temporal discounting research.
 
 # References
