+++ paper/paper.bib
@@ -207,6 +207,14 @@
   doi = {https://doi.org/10.1016/j.jmp.2025.102902}
 }
 
+@article{kinley2024goal,
+  title = {Goal-directed behaviour is associated with decreased temporal discounting},
+  journal = {PsyArXiv},
+  author = {Isaac Kinley and Joseph Oluwasola and Soroush Bagheri and Suzanna Becker},
+  year = {2024},
+  doi = {https://doi.org/10.31234/osf.io/qjcwn}
+}
+
 @article{green1997rate,
   title={Rate of temporal discounting decreases with amount of reward},
   author={Green, Leonard and Myerson, Joel and McFadden, Edward},
