### 2025-03-01 575d461
Add choice rule vignette
