+++ vignettes/choice-rules.Rmd
@@ -1,0 +1,120 @@
+---
+title: "Choice rules"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Choice rules}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+```{r setup}
+library(tempodisco)
+```
+
+## Background
+
+When [modeling binary choice data](https://kinleyid.github.io/tempodisco/articles/modeling-binary-choice-data.html) using [`td_bcnm`](https://kinleyid.github.io/tempodisco/reference/td_bcnm.html), not only the discount function but also the **choice rule** must be specified. Whereas the discount function is parameterized to specify how the subjective value of a future reward decreases with delay, the choice rule is parameterized to specify the probability of choosing either of two rewards whose subjective values are given. To illustrate, `plot(... type = "summary")` visualizes the discount function:
+
+```{r}
+# Load data and fit model
+data("td_bc_single_ptpt")
+mod <- td_bcnm(td_bc_single_ptpt, discount_function = "hyperbolic")
+# Plot the discount curve
+plot(mod, type = "summary", verbose = F)
+```
+
+whereas `plot(... type = "endpoints")` visualizes the choice rule:
+
+```{r}
+# Plot the choice rule
+plot(mod, type = "endpoints", del = 50, val_del = 200, verbose = F)
+```
+
+Just as we can fit a [range of discount functions](https://kinleyid.github.io/tempodisco/articles/comparing-models.html) to the data, we can also fit several choice rules using the `choice_rule` argument to `td_bcnm`. Each of these gives the probability of selecting the immediate reward, $P(I)$, given the subjective values of the immediate and delayed rewards ($V_I$ and $V_D$, respectively, where $V_I$ is simply the face value of the immediate reward and $V_D$ depends on the discount curve) and a "sharpness" parameter $\gamma \geq 0$, where a larger $\gamma$ reflects less "noisy" decision making. Available choice rules are listed below.
+
+## Available choice rules
+
+### Logistic
+
+`choice_rule = "logistic"` (the default) gives the logistic choice rule:
+
+$$
+P(I) = \sigma\left[\gamma \left( V_I - V_D \right) \right]
+$$
+
+where $\sigma$ is the logistic function $x \mapsto (1 + \exp\{-x\})^{-1}$, which the CDF of a logistic distribution with location parameter 0 and scale parameter 1.
+
+### Probit
+
+`choice_rule = "probit"` gives the probit choice rule:
+
+$$
+P(I) = \Phi\left[\gamma \left( V_I - V_D \right) \right]
+$$
+
+where $\Phi$ is the CDF of the standard normal distribution.
+
+### Power
+
+`choice_rule = "power"` gives the power choice rule ([Luce, 1959/2005](https://doi.org/10.1037/14396-000)):
+
+$$
+P(I) = \frac{V_I^\gamma}{V_I^\gamma + V_D^\gamma}
+$$
+
+which is the CDF of a log-logistic distribution with scale parameter 1 and shape parameter $\gamma$.
+
+## Comparing choice rules
+
+Broadly speaking, the logistic and probit choice rules are similar—both are based on a "Fechner model" ([Becker, DeGroot, & Marschak, 1963](https://doi.org/10.1002/bs.3830080106)) in which comparisons between subjective reward values are "noisy":
+
+$$
+P(I) = \Pr(V_I - V_D < \varepsilon )
+$$
+
[…]
