+++ pkgdown.yml
@@ -5,6 +5,7 @@ articles:
   adjusting-amounts: adjusting-amounts.html
   analyzing-study: analyzing-study.html
   area-under-curve: area-under-curve.html
+  choice-rules: choice-rules.html
   comparing-models: comparing-models.html
   custom-discount-functions: custom-discount-functions.html
   drift-diffusion-models: drift-diffusion-models.html
@@ -12,7 +13,7 @@ articles:
   nonsystematic-discounting: nonsystematic-discounting.html
   tempodisco: tempodisco.html
   visualizing-models: visualizing-models.html
-last_built: 2025-02-26T23:24Z
+last_built: 2025-03-01T17:04Z
 urls:
   reference: https://kinleyid.github.io/tempodisco/reference
   article: https://kinleyid.github.io/tempodisco/articles
