### 2025-03-02 b55b700
Add fixed endpoints section to choice rules vignette
