+++ vignettes/choice-rules.Rmd
@@ -25,12 +25,12 @@ When [modeling binary choice data](https://kinleyid.github.io/tempodisco/article
 ```{r}
 # Load data and fit model
 data("td_bc_single_ptpt")
-mod <- td_bcnm(td_bc_single_ptpt, discount_function = "hyperbolic")
+mod <- td_bcnm(td_bc_single_ptpt, discount_function = "exponential")
 # Plot the discount curve
 plot(mod, type = "summary", verbose = F)
 ```
 
-whereas `plot(... type = "endpoints")` visualizes the choice rule:
+whereas `plot(... type = "endpoints")` visualizes the choice rule (i.e., the psychometric curve for a given combination of delay and delayed reward value):
 
 ```{r}
 # Plot the choice rule
@@ -106,7 +106,7 @@ logLiks <- c()
 for (entry in names(plot_legend)) {
   choice_rule <- plot_legend[entry]
   mod <- td_bcnm(td_bc_single_ptpt,
-                 discount_function = "hyperbolic",
+                 discount_function = "exponential",
                  choice_rule = choice_rule)
   logLiks[entry] <- logLik(mod)
   preds <- predict(mod, type = 'response', newdata = newdata)
@@ -118,3 +118,79 @@ legend(0, 1,
 ```
 
 As suggested by the log likelihood measurements shown in the plot above, the choice rule can significantly impact model fit ([Wulff & van den Bos, 2017](https://doi.org/10.1177/0956797616664342)). Therefore it is a good idea to explore multiple choice rules for a given dataset.
+
+## Fixed-endpoint choice rules
+
+As described earlier, for the power choice rule, the probability of choosing an immediate reward of \$0 is 0. In that sense, the left endpoint of the psychometric curve is "fixed" at 0. This is arguably an advantage, if choosing an immediate reward of \$0 should be taken as a failed attention check rather than a reflection of an individual's true preferences (Kinley, Oluwasola & Becker, 2025). Following this logic, we can create choice rules in which both endpoints are fixed, i.e., where the probability of choosing \$0 is 0 and where the probability of choosing \$*x* now over \$*x* at some delay is 1, for all *x*.
+
+This option is available through the `fixed_ends` argument to `td_bcnm`. To see its effect, we can re-fit the models from the plot above with `fixed_ends = TRUE`:
+
+```{r}
+vis_del <- sort(unique(td_bc_single_ptpt$del))[2]
+vis_val_del <- 200
+newdata <- data.frame(del = vis_del,
+                      val_del = vis_val_del,
+                      val_imm = seq(0, vis_val_del, length.out = 1000))
+plot(imm_chosen ~ val_imm, data = subset(td_bc_single_ptpt, del == vis_del),
+     xlim = c(0, vis_val_del), ylim = c(0, 1))
+plot_legend <- c("red" = "logistic",
+                 "blue" = "probit",
+                 "forestgreen" = "power")
+logLiks <- c()
+for (entry in names(plot_legend)) {
+  choice_rule <- plot_legend[entry]
+  mod <- td_bcnm(td_bc_single_ptpt,
+                 discount_function = "exponential",
+                 fixed_ends = TRUE,                 # Fixed endpoints
+                 choice_rule = choice_rule)
+  logLiks[entry] <- logLik(mod)
+  preds <- predict(mod, type = 'response', newdata = newdata)
+  lines(preds ~ newdata$val_imm, col = entry)
+}
+legend(0, 1,
+       fill = names(plot_legend),
+       legend = paste(plot_legend, '; log lik. = ', round(logLiks, 1), sep = ''))
+```
+
[…]
