### 2025-03-02 a78bb99
Add custom choice rules to choice rules vignette
