+++ vignettes/choice-rules.Rmd
@@ -37,36 +37,36 @@ whereas `plot(... type = "endpoints")` visualizes the choice rule (i.e., the psy
 plot(mod, type = "endpoints", del = 50, val_del = 200, verbose = F)
 ```
 
-Just as we can fit a [range of discount functions](https://kinleyid.github.io/tempodisco/articles/comparing-models.html) to the data, we can also fit several choice rules using the `choice_rule` argument to `td_bcnm`. Each of these gives the probability of selecting the immediate reward, $P(I)$, given the subjective values of the immediate and delayed rewards ($V_I$ and $V_D$, respectively, where $V_I$ is simply the face value of the immediate reward and $V_D$ depends on the discount curve) and a "sharpness" parameter $\gamma \geq 0$, where a larger $\gamma$ reflects less "noisy" decision making. Available choice rules are listed below.
+Just as we can fit a [range of discount functions](https://kinleyid.github.io/tempodisco/articles/comparing-models.html) to the data, we can also fit several choice rules using the `choice_rule` argument to `td_bcnm`. Each of these gives the probability of selecting the immediate reward, $P(I)$, given the subjective values of the immediate and delayed rewards ($SV_I$ and $SV_D$, respectively, where $SV_I$ is simply the face value of the immediate reward and $SV_D$ depends on the discount curve) and a "sharpness" parameter $\gamma \geq 0$, where a larger $\gamma$ reflects less "noisy" decision making. Built-in choice rules are listed below.
 
-## Available choice rules
+## Built-in choice rules
 
 ### Logistic
 
 `choice_rule = "logistic"` (the default) gives the logistic choice rule:
 
 $$
-P(I) = \sigma\left[\gamma \left( V_I - V_D \right) \right]
+P(I) = \sigma\left[\gamma \left( SV_I - SV_D \right) \right]
 $$
 
 where $\sigma$ is the logistic function $x \mapsto (1 + \exp\{-x\})^{-1}$, which the CDF of a logistic distribution with location parameter 0 and scale parameter 1.
 
-### Probit
+### Probit (normal)
 
 `choice_rule = "probit"` gives the probit choice rule:
 
 $$
-P(I) = \Phi\left[\gamma \left( V_I - V_D \right) \right]
+P(I) = \Phi\left[\gamma \left( SV_I - SV_D \right) \right]
 $$
 
 where $\Phi$ is the CDF of the standard normal distribution.
 
-### Power
+### Power (log-logistic)
 
 `choice_rule = "power"` gives the power choice rule ([Luce, 1959/2005](https://doi.org/10.1037/14396-000)):
 
 $$
-P(I) = \frac{V_I^\gamma}{V_I^\gamma + V_D^\gamma}
+P(I) = \frac{SV_I^\gamma}{SV_I^\gamma + SV_D^\gamma}
 $$
 
 which is the CDF of a log-logistic distribution with scale parameter 1 and shape parameter $\gamma$.
@@ -76,13 +76,13 @@ which is the CDF of a log-logistic distribution with scale parameter 1 and shape
 Broadly speaking, the logistic and probit choice rules are similar—both are based on a "Fechner model" ([Becker, DeGroot, & Marschak, 1963](https://doi.org/10.1002/bs.3830080106)) in which comparisons between subjective reward values are "noisy":
 
 $$
-P(I) = \Pr(V_I - V_D < \varepsilon )
+P(I) = \Pr(SV_I - SV_D < \varepsilon )
 $$
 
 where the noise, represented by the random variable $\varepsilon$, is assumed to follow either a logistic distribution for the logistic choice rule, or a normal distribution for the probit choice rule. In contrast, the power choice rule could be taken to imply the following Fechner-like model:
 
 $$
-P(I) = \Pr\left( \frac{V_I}{V_D} < \varepsilon \right)
[…]
