+++ vignettes/choice-rules.Rmd
@@ -37,7 +37,7 @@ whereas `plot(... type = "endpoints")` visualizes the choice rule (i.e., the psy
 plot(mod, type = "endpoints", del = 50, val_del = 200, verbose = F)
 ```
 
-Just as we can fit a [range of discount functions](https://kinleyid.github.io/tempodisco/articles/comparing-models.html) to the data, we can also fit several choice rules using the `choice_rule` argument to `td_bcnm`. Each of these gives the probability of selecting the immediate reward, $P(I)$, given the subjective values of the immediate and delayed rewards ($SV_I$ and $SV_D$, respectively, where $SV_I$ is simply the face value of the immediate reward and $SV_D$ depends on the discount curve) and a "sharpness" parameter $\gamma \geq 0$, where a larger $\gamma$ reflects less "noisy" decision making. Built-in choice rules are listed below.
+Just as we can fit a [range of discount functions](https://kinleyid.github.io/tempodisco/articles/comparing-models.html) to the data, we can also fit several choice rules using the `choice_rule` argument to `td_bcnm`. Mathematically, each choice rule gives the probability of selecting the immediate reward, $P(I)$, given the subjective values of the immediate and delayed rewards $SV_I$ and $SV_D$ respectively (where $SV_I$ is simply the face value of the immediate reward and $SV_D$ depends on the discount curve), and a "sharpness" parameter $\gamma \geq 0$, where a larger $\gamma$ reflects less "noisy" decision making. Built-in choice rules are listed below.
 
 ## Built-in choice rules
 
@@ -49,7 +49,7 @@ $$
 P(I) = \sigma\left[\gamma \left( SV_I - SV_D \right) \right]
 $$
 
-where $\sigma$ is the logistic function $x \mapsto (1 + \exp\{-x\})^{-1}$, which the CDF of a logistic distribution with location parameter 0 and scale parameter 1.
+where $\sigma$ is the logistic function $\sigma[x] = (1 + \exp\{-x\})^{-1}$, which is the cumulative distribution function (CDF) of a logistic distribution with location parameter 0 and scale parameter 1.
 
 ### Probit (normal)
 
@@ -69,7 +69,7 @@ $$
 P(I) = \frac{SV_I^\gamma}{SV_I^\gamma + SV_D^\gamma}
 $$
 
-which is the CDF of a log-logistic distribution with scale parameter 1 and shape parameter $\gamma$.
+where the right hand side of the equation is the CDF of a log-logistic distribution with scale parameter 1 and shape parameter $\gamma$.
 
 ## Comparing choice rules
 
@@ -90,7 +90,7 @@ where $\varepsilon$ is now assumed to follow a log-logistic distribution. Mathem
 These theoretical differences between the choice rules have the following practical implications:
 
 1.  In the logistic and probit choice rules, decisions are less stochastic (i.e., $\gamma$ is effectively higher) for larger reward magnitudes, whereas decision stochasticity is constant (i.e., $\gamma$ is effectively constant) for the power choice rule. There is evidence that real human decision making is indeed less stochastic for larger reward values ([Gershman & Bhui, 2020](https://doi.org/10.1038/s41467-020-16852-y)). However, it is not obvious that trying to capture this effect for data with a narrow range of reward magnitudes will improve model fit.
-2.  For the logistic and probit choice rules, $P(I)$ is never exactly 0 or 1. In contrast, for the power choice rule, $P(I) = 0$ when $SV_I = 0$. Thus, for the power choice rule, model fit will be strongly negatively impacted if a participant ever chose an immediate reward of 0. However, such decisions are arguably better used for attention checks than model fitting (cf. [Almog et al., 2023](doi.org/10.1037/pha0000645)).
[…]
