+++ vignettes/modeling-binary-choice-data.Rmd
@@ -20,7 +20,11 @@ library(tempodisco)
 
 ## Classic methods
 
-For binary choice data not explicitly designed to titrate out indifference points (as in an adjusting amount procedure), there are a few analysis options. A common one is the scoring method designed for the Monetary Choice Questionnaire [(Kirby, 1999)](https://doi.org/10.1037//0096-3445.128.1.78):
+For binary choice data not explicitly designed to [titrate out indifference points](https://kinleyid.github.io/tempodisco/articles/adjusting-amounts.html) (as in an adjusting amount procedure), there are a few widely-used traditional scoring methods to quantify discounting.
+
+### Kirby scoring
+
+One scoring method is the one designed for the Monetary Choice Questionnaire [(Kirby, 1999)](https://doi.org/10.1037//0096-3445.128.1.78):
 
 ```{r}
 data("td_bc_single_ptpt")
@@ -28,14 +32,16 @@ mod <- kirby_score(td_bc_single_ptpt)
 print(mod)
 ```
 
-Although this method computes $k$ values according to the hyperbolic discount function, in principle it's possible to use the exponential discount function:
+Although this method computes $k$ values according to the hyperbolic discount function, in principle it's possible to use the exponential discount function (thought this is not an established practice and should be considered an experimental feature of `tempodisco`):
 
 ```{r}
 mod <- kirby_score(td_bc_single_ptpt, discount_function = 'exponential')
 print(mod)
 ```
 
-Another option is to use the logistic regression method of [Wileyto et al. (2004)](https://doi.org/10.3758/BF03195548), where we can solve for the $k$ value of the hyperbolic discount function in terms of the regression coefficients:
+### Wileyto scoring
+
+It is also possible to use the logistic regression method of [Wileyto et al. (2004)](https://doi.org/10.3758/BF03195548), where we can solve for the $k$ value of the hyperbolic discount function in terms of the regression coefficients:
 
 ```{r}
 mod <- wileyto_score(td_bc_single_ptpt)
@@ -44,7 +50,9 @@ print(mod)
 
 ## Newer methods
 
-The [Wileyto et al. (2004)](https://doi.org/10.3758/BF03195548) approach turns out to be possible for other discount functions as well:
+### Linear models
+
+The [Wileyto et al. (2004)](https://doi.org/10.3758/BF03195548) approach turns out to be possible for other discount functions as well ([Kinley, Oluwasola & Becker, 2025](https://doi.org/10.1016/j.jmp.2025.102902).):
 
 ```{r child="../man/fragments/linear-models.Rmd"}
 ```
@@ -56,7 +64,9 @@ mod <- td_bclm(td_bc_single_ptpt, model = 'all')
 print(mod)
 ```
 
-To explore a wider range of discount functions, we can fit a nonlinear model by calling `td_bcnm`. The full list is as follows:
+### Nonlinear models
+
+To explore a wider range of discount functions, we can fit a nonlinear model by calling `td_bcnm`. The full list of built-in discount functions is as follows:
 
 ```{r child="../man/fragments/predefined-discount-functions.Rmd"}
 ```
@@ -66,7 +76,9 @@ mod <- td_bcnm(td_bc_single_ptpt, discount_function = 'all')
 print(mod)
 ```
 
-Several additional arguments can be used to customize the model. For example, we can use different choice rules---the "logistic" choice rule is the default, but the "probit" and "power" choice rules are also available (see [Wulff and Van den Bos, 2018](https://doi.org/10.1177/0956797616664342), for explanations of these):
+#### Choice rules
+
[…]
