### 2025-03-07 b876864
Add option for custom max_abs_drift_par_starts
