### 2025-03-07 75c43bd
Add "power" option for kirby_score
