### 2025-03-24 2b4d777
Increment version number to 1.0.0
