### 2025-04-03 0c0d1f0
Updates for CRAN
