+++ man/td_bc_study.Rd
@@ -8,7 +8,7 @@
 Data from 421 participants, who each made 70 binary choices. Along with the columns required by \code{td_bcnm}, the reaction time (\code{rt}) in seconds is recorded. Participants are identified by the alphnumeric code in the \code{id} column.
 }
 \references{
-\url{doi.org/10.31234/osf.io/y2fdh}
+\url{https://doi.org/10.31234/osf.io/y2fdh}
 }
 \author{
 Isaac Kinley \email{isaac.kinley@gmail.com}
