### 2025-04-03 #7472 @skanwal
@kinleyid - can you please double check reference dois based on the suggestions from the bot [above](https://github.com/openjournals/joss-reviews/issues/7472#issuecomment-2777207297)?
