+++ paper/paper.bib
@@ -132,6 +132,7 @@
   number={11},
   pages={1176--1186},
   year={2019},
+  doi={10.1001/jamapsychiatry.2019.2102},
   publisher={American Medical Association}
 }
 
@@ -163,6 +164,7 @@
   number={4},
   pages={1608--1620},
   year={2016},
+  doi={10.3758/s13428-015-0672-2},
   publisher={Springer}
 }
 
@@ -173,6 +175,7 @@
   volume={39},
   pages={293--304},
   year={2016},
+  doi={10.1007/s40614-016-0070-9},
   publisher={Springer}
 }
 
@@ -193,6 +196,7 @@
   author = {Jonathan E. Friedel},
   year = {2024},
   note = {R package version 1.0.3},
+  doi={10.32614/CRAN.package.discAUC},
   url = {https://CRAN.R-project.org/package=discAUC},
 }
 
@@ -204,7 +208,7 @@
   pages = {102902},
   year = {2025},
   issn = {0022-2496},
-  doi = {https://doi.org/10.1016/j.jmp.2025.102902}
+  doi = {10.1016/j.jmp.2025.102902}
 }
 
 @article{kinley2024goal,
@@ -212,7 +216,7 @@
   journal = {PsyArXiv},
   author = {Isaac Kinley and Joseph Oluwasola and Soroush Bagheri and Suzanna Becker},
   year = {2024},
-  doi = {https://doi.org/10.31234/osf.io/qjcwn}
+  doi = {10.31234/osf.io/qjcwn}
 }
 
 @article{green1997rate,
@@ -222,5 +226,6 @@
   volume={25},
   pages={715--723},
   year={1997},
+  doi={10.3758/BF03211314},
   publisher={Springer}
 }
