### 2025-04-05 97ab192
"mod" -> "object" in generics docs
