+++ man/logLik.td_ipm.Rd
@@ -7,9 +7,9 @@
 \method{logLik}{td_ipm}(object, ...)
 }
 \arguments{
-\item{...}{Additional arguments currently not used.}
+\item{object}{An object of class \code{td_ipm}}
 
-\item{mod}{An object of class \code{td_ipm}}
+\item{...}{Additional arguments currently not used.}
 }
 \description{
 Compute log-likelihood for a temporal discounting indifference point model.
