+++ man/deviance.td_bcnm.Rd
@@ -7,9 +7,9 @@
 \method{deviance}{td_bcnm}(object, ...)
 }
 \arguments{
-\item{...}{Additional arguments currently not used.}
+\item{object}{An object of class \code{td_bcnm}.}
 
-\item{mod}{An object of class \code{td_bcnm}.}
+\item{...}{Additional arguments currently not used.}
 }
 \description{
 Compute deviance for a temporal discounting binary choice model.
