### 2025-04-06 #7472 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
