### 2025-04-07 #7472 @crvernon
@kinleyid - Please address the following before I accept this for publication:

In your archive:

- [ ] The name of your Zenodo archive must match the title of your paper exactly.  Please edit your archive metadata to fix this.  No new release required.

In your paper:

- [ ] LINE 100:  The "b" in "bayesian" should be capitalized.  You can do this by placing curly brackets around the characters you wish to maintain formatting of in the bib file.

Let me know when these have been addressed.  Thanks!
