+++ paper/paper.bib
@@ -34,7 +34,7 @@
 }
 
 @article{franck2015accurate,
-  title={Accurate characterization of delay discounting: A multiple model approach using approximate Bayesian model selection and a unified discounting measure},
+  title={Accurate characterization of delay discounting: A multiple model approach using approximate {B}ayesian model selection and a unified discounting measure},
   author={Franck, Christopher T and Koffarnus, Mikhail N and House, Leanna L and Bickel, Warren K},
   journal={Journal of the experimental analysis of behavior},
   volume={103},
@@ -157,7 +157,7 @@
 }
 
 @article{vincent2016hierarchical,
-  title={Hierarchical Bayesian estimation and hypothesis testing for delay discounting tasks},
+  title={Hierarchical {B}ayesian estimation and hypothesis testing for delay discounting tasks},
   author={Vincent, Benjamin T},
   journal={Behavior research methods},
   volume={48},
